### 2023-12-15 1f6cc65
Update references.bib
