+++ paper/references.bib
@@ -49,7 +49,8 @@
   number={2},
   pages={e1006830},
   year={2019},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pcbi.1006830}
 }
 
 @article{10.1093/bioinformatics/bti431,
