+++ src/strengths/engine_collection.py
@@ -1,6 +1,7 @@
 import platform
 import strengths
 import ctypes
+import pathlib
 
 from strengths.librdengine import LibRDEngine
 
@@ -10,19 +11,20 @@ def _get_strengths_path():
             path += i
         return path
 
-def _get_os_specific_engine_lib_name() :
-    os = platform.system()
-    if   os == "Windows" :
-        return "engine_w.dll"
-    elif os == "Linux" :
-        return "engine_l.so"
-    elif os == "Darwin" :
-        return "engine_m.so"
-    else :
-        raise OSError("It seems that we havent planed any precompiled engine " +
-                      "libaray for your operating system. You can try to " +
-                      "build the engine dynamic library from sources, and then " +
-                      "use it through a LibRDEngineRefrence object.")
+def _get_engine_path() :
+    # Assumes the engine library binary file is the 
+    # only non directory item in the strengths_engine 
+    # directory.
+
+    strnpath = _get_strengths_path()
+    sedirpath = strnpath + "/engines/strengths_engine"
+    path = pathlib.Path(sedirpath)
+    content = path.iterdir()
+    files = []
+    for i in content :
+        if not i.is_dir() : 
+            files.append(i)
+    return str(files[0])
 
 def gillespie_engine():
     """
@@ -33,8 +35,7 @@ def gillespie_engine():
     # .. [#Bernstein2005] Bernstein, D. (2005). Simulating mesoscopic reaction-diffusion systems using the Gillespie algorithm. Physical Review E, 71(4), Article 041103. https://doi.org/10.1103/PhysRevE.71.041103
     # .. [#Gillespie1977] Gillespie, D. T. (1977). Exact stochastic simulation of coupled chemical reactions. The Journal of Physical Chemistry, 81(25), 2340-2361. https://doi.org/10.1021/j100540a008
     
-    path = _get_strengths_path()+"/engines/strengths_engine/"+_get_os_specific_engine_lib_name()
-    
+    path = _get_engine_path()
     return LibRDEngine(
         ctypes.CDLL(path),
         option="gillespie",
@@ -51,8 +52,7 @@ def tauleap_engine():
     # .. [#Bernstein2005] Bernstein, D. (2005). Simulating mesoscopic reaction-diffusion systems using the Gillespie algorithm. Physical Review E, 71(4), Article 041103. https://doi.org/10.1103/PhysRevE.71.041103
     # .. [#Gillespie2001] Gillespie, D. T. (2001). Approximate accelerated stochastic simulation of chemically reacting systems. The Journal of Chemical Physics, 115(4), 1716-1733. https://doi.org/10.1063/1.1378322
     
-    path = _get_strengths_path()+"/engines/strengths_engine/"+_get_os_specific_engine_lib_name()
-
+    path = _get_engine_path()
     return LibRDEngine(
         ctypes.CDLL(path),
         option = "tauleap",
@@ -68,7 +68,7 @@ def euler_engine():
     # references :
     # .. [#Bernstein2005] Bernstein, D. (2005). Simulating mesoscopic reaction-diffusion systems using the Gillespie algorithm. Physical Review E, 71(4), Article 041103. https://doi.org/10.1103/PhysRevE.71.041103
 
-    path = _get_strengths_path()+"/engines/strengths_engine/"+_get_os_specific_engine_lib_name()
+    path = _get_engine_path()
     return LibRDEngine(
         ctypes.CDLL(path),
         option = "euler",
