### 2023-12-21 0afa472
Update SimulationAlgorithmGraphBase.hpp
