### 2023-12-21 9362c80
Update SimulationAlgorithmGraphBase.hpp
