### 2023-12-21 a78ff0a
Improving build and distribution (#1)
* Update MANIFEST.in

adding missing engine sources

* Update engine_collection.py

* Update engine.cpp

* Create setup.py

* Update engine.cpp

* Update engine.cpp

* Update engine.cpp

* Update SimulationAlgorithm3DBase.hpp

* Update SimulationAlgorithmGraphBase.hpp

* Update engine.cpp

* Update SimulationAlgorithmGraphBase.hpp

* Update engine.cpp

* Update engine.cpp

* Update engine.cpp

* Update pyproject.toml

increment version
