+++ MANIFEST.in
@@ -3,4 +3,7 @@ include src/strengths/engines/strengths_engine/src/Euler3D.hpp
 include src/strengths/engines/strengths_engine/src/Gillespie3D.hpp
 include src/strengths/engines/strengths_engine/src/SimulationAlgorithm3DBase.hpp
 include src/strengths/engines/strengths_engine/src/TauLeap3D.hpp
-include src/strengths/engines/strengths_engine/engine_w.dll
+include src/strengths/engines/strengths_engine/src/EulerGraph.hpp
+include src/strengths/engines/strengths_engine/src/GillespieGraph.hpp
+include src/strengths/engines/strengths_engine/src/SimulationAlgorithmGraphBase.hpp
+include src/strengths/engines/strengths_engine/src/TauLeapGraph.hpp
