+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "strengths"
-version = "0.0.8"
+version = "0.0.9"
 authors = [
 	{name = "Thibault Fillion"},
 	{name = "Francesco Piazza"}]
