+++ setup.py
@@ -7,6 +7,7 @@ setup(
              sources = ["src/strengths/engines/strengths_engine/src/engine.cpp"],
              include_dirs = ["src/strengths/engines/strengths_engine/src"],
              define_macros = [("CPYEMVER", None)],
+             extra_compile_args = ["-std=c++11"],
              export_symbols = [
                 "Initialize3D",
                 "InitializeGraph",
