+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "strengths"
-version = "0.0.9"
+version = "0.0.10"
 authors = [
 	{name = "Thibault Fillion"},
 	{name = "Francesco Piazza"}]
