### 2024-01-11 #6153 @Kevin-Mattheus-Moerman
@ThibaultFillion thanks for this submission. While I look for a handling editor please can you work on the following:

- [ ] Please check the above potentially missing/invalid DOIs ☝️ and amend your .bib file where needed. You can call `@editorialbot check references` to check them again.
- [ ] Please expand your README to include more information as is customary e.g. links to documentation/testing, an installation/ getting started section, contributing guidelines (or a link to them), and a link to the license.
