### 2024-01-12 921f460
Create building_strengths_from_source.rst
