### 2024-01-12 df13c3d
Update using_strengths.rst
