### 2024-01-12 418e102
Update using_strengths.rst
