### 2024-01-13 #6153 @Nikoleta-v3
Great, thank you for fixing the DOI @ThibaultFillion. When you address the comments regarding the README, please ping me here 😄 Also, could you include the commit hash when you do?
