### 2024-01-15 b7a3a44
Update engine_collection.py
Adding a guiding error message for when the 
engine shared library is missing.
