### 2024-01-15 0b256b2
Update README.rst
Adding the "Testing" section
