### 2024-01-16 57133ea
Create community_and_contribution_guidelines.rst
