### 2024-01-16 2dc23fb
Update community_and_contribution_guidelines.rst
