### 2024-01-16 058c5c9
Update README.rst
adding content to the "Source code and contribution" section
