### 2024-01-16 92d32db
Update community_and_contribution_guidelines.rst
