### 2024-01-16 441b72d
Update community_and_contribution_guidelines.rst
