### 2024-01-16 563d8f4
Update community_and_contribution_guidelines.rst
