### 2024-01-16 0ec46e9
Update community_and_contribution_guidelines.rst
