### 2024-01-16 68d0741
Update community_and_contribution_guidelines.rst
