### 2024-01-16 efeff9b
Update community_and_contribution_guidelines.rst
