+++ community_and_contribution_guidelines.rst
@@ -5,7 +5,7 @@ Welcome and thank you for your interest in the Strengths package.
 Please, be respectful to each other, and maintain a safe, healthy and pleasant work environment for everyone.
 While the project's repository is hosted on GitHub, please make sure to respect 
 `GitHub's code of conduct <https://docs.github.com/en/site-policy/github-terms/github-community-code-of-conduct>`_.
-More specific gidelines for the project may be added or linked here in the future.
+Specific gidelines for the project may be added or linked here in the future.
 
 Contributing
 ============
