+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "strengths"
-version = "0.0.10"
+version = "0.0.11"
 authors = [
 	{name = "Thibault Fillion"},
 	{name = "Francesco Piazza"}]
