### 2024-01-16 6a4ae34
completing readme + contribution + documentation (#3)
* Create building_strengths_from_source.rst

* Update using_strengths.rst

* Update using_strengths.rst

* Update README.rst

* Update README.rst

* Update README.rst

* Update README.rst

* Update README.rst

* Update README.rst

* Update engine_collection.py

Adding a guiding error message for when the 
engine shared library is missing.

* Update engine_collection.py

* Update engine_collection.py

* Update engine_collection.py

* Update README.rst

Adding the "Testing" section

* Create community_and_contribution_guidelines.rst

* Update
[…]
