### 2024-01-16 #6153 @ThibaultFillion
Sorry for the delay. I tried to expand the README file as requested.
The corresponding commit hash is 6a4ae34 (i think).
