### 2024-01-22 #6153 @Nikoleta-v3
Thank you @ThibaultFillion for your work on the Readme! I just have two questions:

1. Regarding the contribution guidelines, is there a particular reason why you are referring to this [copy](https://raw.githubusercontent.com/ThibaultFillion/strengths/main/community_and_contribution_guidelines.rst) of the file  instead of this [one](https://github.com/ThibaultFillion/strengths/blob/main/community_and_contribution_guidelines.rst)?

2. Regarding your test setup, it's great that you have implemented tests. As you point out, one has to run the command

```
python run_all_tests.py
```

Are you familiar with the package [pytest](https://docs.pytest.org/en/7.4.x/)? I recommend pytest because it works similarly to your run file. It runs all files that have "test" in the name and each function that also has "test" in its name, providing more informative output.

This is what it looks like when I run `pytest` in the tests folder. You can also run `pytest tests` from the root of the directory.

I'm not saying you should use `pytest`; it's just a recommendation. If not, could you state something like, "If the command runs and you receive no output, then the tests have passed." after the `python run_all_tests.py`.

At this point, could you also please check out this page: https://reviewers.joss.theoj.org/lookup and suggest at least three reviewers? 😄
