### 2024-02-10 #6153 @ThibaultFillion
@Nikoleta-v3 Sorry again for the delay. To answer your questions :

1. I have made a mistake here, linking the github page would be better.
I will edit the links, thank you for telling me.

2. Thank you for recommending pytest,
it would indeed work great with the current implementation.
For now, i will just go for the second option,
but i might definitly switch for pytest in the future.
Thank you again for the advice.

For the reviewers, we recommend :
* ggmirandac
* lorenzo-rovigatti
* ASinanSaglam (not on the list)
