### 2024-02-19 b9910b3
Update rdnetwork.py
