### 2024-02-19 ab3e3fb
Update run_all_tests.py
adding test_kinetics
