### 2024-02-19 2571909
Update README.rst
fixing urls
