### 2024-02-19 ef88adc
Create strengths_and_scipy.rst
