### 2024-02-19 fa351b3
Update strengths_and_scipy.rst
