### 2024-02-19 3316469
Update using_strengths.rst
adding strengths_and_scipy.rst to the toc
