### 2024-02-19 e86c1d2
Global improvements (#4)
* Update kinetics.py

improved for compatibility with both rdgridspace and rdgraphspace

* Update rdnetwork.py

* Create test_kinetics.py

* Update run_all_tests.py

adding test_kinetics

* Update run_all_tests.py

* Update README.rst

fixing urls

* Update pyproject.toml

incrementing version

* Create strengths_and_scipy.rst

* Rename strengths_and_scipy.rst to documentation/strengths_and_scipy.rst

fixing file location

* Update strengths_and_scipy.rst

* Update using_strengths.rst

adding strengths_and_scipy.rst to the toc
