+++ tests/test_kinetics.py
@@ -1,0 +1,447 @@
+import sys
+sys.path.append("../src")
+from strengths import *
+from strengths.constants import avogadro_number
+import strengths.kinetics as strnkin
+import numpy as np
+
+# test kintetics module
+
+##################################
+# test reaction rate
+# should be done both for rdnetworks and graphs
+
+def test_compute_reaction_rates_dimensions_for_different_reaction_orders_with_default_units_default_units(spacetype) :
+    #checks that rates are in molecule/s, no mater the reaction order
+ 
+    system_dict = {
+        "network" : {
+            "species" : [
+                    {"label" : "A"},
+                    {"label" : "B"},
+                    {"label" : "C"},
+                    {"label" : "D"},
+                    {"label" : "E"},
+                    {"label" : "F"}
+                ],
+            "reactions" : [
+                    {"eq" : " -> ", "kf" : 2,  "kr" : 3},
+                    {"eq" : "0 A + 0 B  -> 0 C", "kf" : 2,  "kr" : 3},
+                    {"eq" : "A -> B", "kf" : 2,  "kr" : 3},
+                    {"eq" : "A + B -> C", "kf" : 2,  "kr" : 3},
+                    {"eq" : "A + B + C-> C + D + E + F", "kf" : 2,  "kr" : 3},
+                    {"eq" : "A + B -> C", "kf" : 2,  "kr" : 3},
+                    {"eq" : "2 C + 4 B -> 6 C + 0 E + 2 F", "kf" : 2,  "kr" : 3}
+                ]
+            }
+        }
+    if spacetype=="graph" :
+        system_dict["space"] = {
+            "type" : "graph",
+            "nodes" : [{}],
+            "edges" : []}
+
+    system = rdsystem_from_dict(system_dict)
+
+    state = UnitArray([1,1,1,1,1,1], "molecule")
+
+    for i in range(len(system.network.reactions)) :
+        rates = strnkin.compute_reaction_rates(
+            system = system,
+            reaction = system.network.reactions[i],
+            position = 0,
+            state = state,
+            units_system = UnitsSystem())
+        for rate in rates :
+            assert rate.units.dim == UnitsDimensions(quantity=1, space=0, time=-1)
+
+def test_compute_reaction_rates_value(spacetype) :
+    #checks that rates computed values are correct
+    kf = 12
+    kr = 0.5
+    V = 1.23e-3
+
+    A = 5
+    B = 90
+    C = 1
+    D = 548
+    E = 0.33
+    F = 1000
+    
+    NA = avogadro_number()
+    state = UnitArray([A, B, C, D, E, F], "mol")    
+
+    system_dict = {
+        "network" : {
+            "species" : [
+                    {"label" : "A"},
+                    {"label" : "B"},
+                    {"label" : "C"},
+                    {"label" : "D"},
+                    {"label" : "E"},
+                    {"label" : "F"}
+                ],
+            "reactions" : [
+                    {"eq" : "2 C + 4 B + 1 A -> 6 D + 0 E + 3 F", "kf" : str(kf)+" µm18/molecule6/min",  "kr" : kr}
+                ]
+            }
+        }
+
+    if spacetype=="graph" :
+        system_dict["space"] = {
+            "type" : "graph",
+            "nodes" : [{"vol" : str(V)+" mL"}],
+            "edges" : []}
+    else :
+        system_dict["space"] = {"cell_vol" : str(V)+" mL"}
+        
+    system = rdsystem_from_dict(system_dict)
+
+    V *= 1e12
+    rf = kf/60 * V * NA*A/V * (NA*C/V)**2 * (NA*B/V)**4 # molecule/s
[…]
