+++ src/strengths/kinetics.py
@@ -2,9 +2,12 @@ import numpy as np
 import json
 import strengths.value_processing as valproc
 from strengths.units import *
+from strengths.rdnetwork import Reaction
+from strengths import value_processing as valproc
+from strengths import RDGraphSpace, RDGridSpace
 
 """
-Module that computation functions
+Module that implements computation functions
 associated with reaction diffusion systems.
 """
 
@@ -19,7 +22,9 @@ def compute_reaction_rates(system,
     
     :param system: system in which the reaction occurs.
     :type system: RDSystem
-    :param reaction: reaction of interest.
+    :param reaction: reaction of interest. If a Reaction object is given,
+        if will be used directly. if an index or a lable is given, the reaction will be
+        retrived from the system's network.
     :type reaction: Reaction, number or string
     :param position: position of the cell in which the rates should be computed
     :type position: number, tuple or coord like
@@ -34,27 +39,36 @@ def compute_reaction_rates(system,
     if isnone(state) :
         state = system.state
     
-    reaction_index = system.network.get_reaction_index(reaction)
-    
+    if type(reaction) != Reaction :
+        reaction_index = system.network.get_reaction_index(reaction)
+        if reaction_index is None :
+            raise ValueError("No index match the given reaction.")
+        if isarray(reaction_index) :
+            raise ValueError("Ambiguous reaction argument have multiple index. Try passing a reaction index or label instead?")
+        reaction = system.network.reactions[reaction_index]
+
     position_index = system.get_cell_index(position)
     
-    r = system.network.reactions[reaction_index]
     
-    ssto = r.ssto(system.network.species_labels())
-    psto = r.psto(system.network.species_labels())
+    ssto = reaction.ssto(system.network.species_labels())
+    psto = reaction.psto(system.network.species_labels())
     
-    volume = system.space.cell_vol.copy()
+    volume = system.space.get_cell_vol_array().get_at(position_index)
     
-    if not r.environments==None and not system.network.environments[system.cell_env[position_index]] in r.environments :
-        rates = [ unitValue(0, Units(sys=units_system, dim=r.kf_units_dimensions)),
-                  unitValue(0, Units(sys=units_system, dim=r.kr_units_dimensions)) ]
-    else :
-        rf, rr = r.kf * volume, r.kr * volume
-        for i in range(system.network.nspecies()) :
-            state_index = system.get_state_index(species=i, position=position_index)
-            rf *= (state.get_at(state_index)/volume)**ssto[i]
-            rr *= (state.get_at(state_index)/volume)**psto[i]
-        return rf.convert(units_system), rr.convert(units_system)
+    if reaction.environments is not None :
+        environment_index = system.space.get_cell_env_array()[position_index]
+        environment_label = system.network.environments[environment_index]
+        if not environment_label in reaction.environments :
+            return [ UnitValue(0, Units(sys=units_system, dim=UnitsDimensions(quantity=1, space=0, time=-1))),
+                     UnitValue(0, Units(sys=units_system, dim=UnitsDimensions(quantity=1, space=0, time=-1))) ]
+
+    rf = reaction.kf * volume
+    rr = reaction.kr * volume
+    for i in range(system.network.nspecies()) :
+        state_index = system.get_state_index(species=i, position=position_index)
+        rf *= (state.get_at(state_index)/volume)**ssto[i]
+        rr *= (state.get_at(state_index)/volume)**psto[i]
+    return rf.convert(units_system), rr.convert(units_system)
 
 
 def compute_diffusion_rates(system, 
@@ -66,9 +80,8 @@ def compute_diffusion_rates(system,
     """
[…]
