### 2024-03-03 #6153 @Wrzlprmft
> 👋🏻 @Wrzlprmft […] would any of you be willing to review this submission for JOSS?

Generally yes, but I am very busy until the mid of April, so I won’t be able to even start the review before then.

At a very first glance, the paper seems rather long (for JOSS’s standards) and contains things like installation instructions that do not need to be there.
