### 2024-03-03 #6153 @parikshitbajpai
> 👋🏻  @Wrzlprmft @parikshitbajpai would any of you be willing to review this submission for JOSS?
>
>
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
>
>
> The submission I'd like you to review is titled: "STReNGTHS, a python package to model and simulate reaction-diffusion systems". You can find more information at the top of this Github issue 🆙
>
>
>
> Please let me know if you're available 😄 Thank you!

Yes, I can review this paper.
