### 2024-03-07 #6153 @Nikoleta-v3
> Generally yes, but I am very busy until the mid of April, so I won’t be able to even start the review before then.
>
> At a very first glance, the paper seems rather long (for JOSS’s standards) and contains things like installation instructions that do not need to be there.

@Wrzlprmft, thanks for getting back to me! 🙏🏻 🚀 No problem at all; I will try to find another reviewer. If I can't find one by mid-April, I might ping you again.

Appreciate your feedback! You're right about the length of the paper. Once we get into the review process, I'll make sure to give the author some pointers on that.
