### 2024-03-07 #6153 @Nikoleta-v3
@corettialessandro @shivupa would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "STReNGTHS, a python package to model and simulate reaction-diffusion systems". You can find more information at the top of this Github issue 🆙
