### 2024-03-16 #6495 @editorialbot
[REVIEW]: STReNGTHS, a python package to model and simulate reaction-diffusion systems
**Submitting author:** @ThibaultFillion (<a href="http://orcid.org/0009-0001-6738-4825">Thibault Fillion</a>)
**Repository:** https://github.com/ThibaultFillion/strengths
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.8
