### 2024-03-16 9de9026
Fix typos in README
