### 2024-03-17 0d3f5c2
Update Readme
Allows install command to be copyable
