### 2024-03-17 ThibaultFillion/strengths#5 @parikshitbajpai
Typos in Readme
While reviewing openjournals/joss-reviews#6495, I found a few typos in the readme.

Some of these are:
- Tri-dimetionnal should be tridimensional
- `pip install strengths` command was quoted text instead of code format.
-  "strengths" and "python" used in several places instead of "Strengths" and "Python respectively".

These are minor changes that in no way affect the review but bringing them to your notice since I spotted them.
