### 2024-03-17 ThibaultFillion/strengths#6 @parikshitbajpai
Clarify installation requirements
During my review on openjournals/joss-reviews#6495, I followed the instruction to install Strengths using pip but forgot to install requirements using `pip install -r requirements.txt`. This meant the tests failed when I ran them the first time. This can easily happen when using separate condo / virtual environments to avoid dependency conflicts between various codes.

I suggest either adding the installing requirements step to the instructions in readme or adding dependencies to `setuptools` configuration file so that the required packages get installed alongside strengths.
