### 2024-03-17 f58e797
Fix typos in README and add dependencies to configuration. (#7)
* Fix typos in README

* Update configuration file to allow dependencies to install with `pip`

* Update Readme

Allows install command to be copyable
