+++ README.rst
@@ -4,21 +4,23 @@ The Strengths Package
 Introduction
 ------------
 
-Strengths is a modeling and simulation tool for reaction diffusion systems, interfaced as a python package.
-It stands for "Simulation and modeling Tool for Reaction diffusion Networks in Graphs and Tri-dimentionnal Heterogenous Systems".
-The design of reaction-diffusion systems if facilitated by the use of dictionary as a way to define most of the key Objects.
+Strengths is a modeling and simulation tool for reaction diffusion systems, interfaced as a Python package.
+It stands for "Simulation and modeling Tool for Reaction diffusion Networks in Graphs and Tridimensional Heterogenous Systems".
+The design of reaction-diffusion systems is facilitated by the use of dictionary as a way to define most of the key objects.
 
 Installing and using Strengths
 ------------------------------
 
 The package can be installed from the Python Package Index, with
 
+.. code:: bash
+
   pip install strengths
 
-Alternatively, you can build strengths from source. More details on how to do so are given in the documentation, in the
-`"Building strengths from source" section <https://strengths.readthedocs.io/en/latest/building_strengths_from_source.html>`_.
+Alternatively, you can build Strengths from source. More details on how to do so are given in the documentation, in the
+`"Building Strengths from source" section <https://strengths.readthedocs.io/en/latest/building_strengths_from_source.html>`_.
 
-Once it is successfully installed, you should be able to import it in python.
+Once it is successfully installed, you should be able to import it in Python.
 Below is a short example, featuring the simulation of the trajectory for a simple
 reaction system :
 
@@ -52,7 +54,7 @@ reaction system :
   # plotting the trajectories of A and B
   strnplt.plot_trajectory(trajectory, ["A", "B"])
 
-More examples, using more advanced features - especially diffusion -, are available in the documentation, in the `"Using strengths" section <https://strengths.readthedocs.io/en/latest/using_strengths.html>`_.
+More examples, using more advanced features - especially diffusion -, are available in the documentation, in the `"Using Strengths" section <https://strengths.readthedocs.io/en/latest/using_strengths.html>`_.
 
 Documentation
 -------------
@@ -67,7 +69,7 @@ methods and attributes are covered in detail.
 Source code and contribution
 ----------------------------
 
-Strengths have a `repository <https://github.com/ThibaultFillion/strengths/tree/main>`_ hosted on GitHub.
+Strengths has a `repository <https://github.com/ThibaultFillion/strengths/tree/main>`_ hosted on GitHub.
 
 If you wish to contribute to the package,
 whether by giving your feedback, reporting bugs or errors,
@@ -77,7 +79,7 @@ please refer to the project's `community and contribution guidelines <https://gi
 Testing
 -------
 
-Running "run_all_tests.py" in the "tests" directory will execute at once all unit tests for the package.
+Running :code:`run_all_tests.py` in the `tests` directory will execute at once all unit tests for the package.
 
 Licence
 -------
