### 2024-03-17 ThibaultFillion/strengths#5 @ThibaultFillion
Thank you for pointing out those mistakes,
i will do a global check for typos in the repository (especially the documentation)
and try to be more careful with the spelling.
