### 2024-03-17 305b3ba
Update README.rst
Adding dependencies requirements to the install section
+ colon syntax fix
