### 2024-03-17 a011b3a
Update apiref.rst
adding caps to section titles
