### 2024-03-17 61a98e3
Update apiref.rst
removing spaces before colons
