### 2024-03-17 d4a5e84
Thibault fillion patch 1 (#8)
* Update README.rst

Adding dependencies requirements to the install section
+ colon syntax fix

* Update apiref.rst

adding caps to section titles

* Update apiref.rst

removing spaces before colons
