### 2024-03-17 ThibaultFillion/strengths#6 @ThibaultFillion
Sorry for the inconvenience, thank you for telling me.
I added instructions to the readme file.
