### 2024-03-17 d057e05
Update installing_and_importing_strengths.rst
making use of hyperlinks
