+++ documentation/installing_and_importing_strengths.rst
@@ -1,7 +1,7 @@
 installing and importing Strengths
 ==================================
 
-The Strengths package is available from the Python Package Index PyPI [#pypi]_ (https://pypi.org/project/strengths/).
+The Strengths package is available from the Python Package Index `PyPI <https://pypi.org/project/strengths/>`_.
 It can be installed with
 
 .. code:: bash
@@ -15,9 +15,4 @@ Once this is done, you should be able to import strengths's core features with
   import strengths
 
 Please look at the dependencies section to learn about packages that you need to have installed
-to have strengths working.
-
-References
-----------
-
-.. [#pypi] Python Software Foundation. the Python Package Index, PyPI. (consulted of september 05, 2023). https://pypi.org/
+to have strengths working, especially `NumPy <https://numpy.org>`_.
