### 2024-03-17 f755691
Update installing_and_importing_strengths.rst
adding caps and colons
