### 2024-03-17 1524931
Update dependencies.rst
hyperlinks seems more appropriate in this context
