### 2024-03-17 9b0d493
Update dependencies.rst
