### 2024-03-17 b32fa3e
Update strengths_and_scipy.rst
