### 2024-03-17 ecf0801
Update setting_up_initial_conditions.rst
