+++ documentation/setting_up_initial_conditions.rst
@@ -4,10 +4,10 @@ Setting up initial conditions
 Setting up the initial state
 ----------------------------
 
-let us consider the following system
+Let us consider the following system
 with the reversible reaction A -> B.
 
-*system.json* :
+*system.json*:
 
 .. code:: json
 
@@ -21,21 +21,23 @@ with the reversible reaction A -> B.
     },
   }
 
-once loaded as a system (a DRSystem object),
-it will have an initial state, in which the quantity of each species in each cell
-will be determined based on the species density and the cell volume.
+Once loaded as a system (a DRSystem object),
+it will has an initial state in which the quantity of each species in each cell
+is determined based on the species density and the cell volume.
 
 However, it is also possible to explicitly set the initial state of the system.
-This can be set directly in the JSON/dictionary or done after the system have been loaded.
+This can be done directly in the JSON file/dictionary or with the RDSystem instance.
 
-To set the state explicitly, one must set the "state" attribute of the system JSON/dictionary,
-or the state argument of the constructor.
+To set the state explicitly, one must set the *"state"* attribute of the system JSON/dictionary,
+the *state* argument of the RDSystem constructor or the
+*state* attribute of the RDSystem object.
 
-We left the space to default, so w=d=h=1.
-the system state will thus be [quantity of A in cell 0, quantity of B in cell 0].
-let us say we initially want 5 nmol of A and 3 nmol of B in the unique cell of the system.
+The *"space*" attribute have been left to default in the example above.
+As a consequence, the dimensions of the reaction diffusion space are w=d=h=1 cell. 
+The system state is thus: [quantity of A in cell 0, quantity of B in cell 0].
+Let us say we initially want 5 nmol of A and 3 nmol of B in the unique cell of the system.
 
-using the JSON/dict script, it will be :
+Using the JSON/dict script, it will be:
 
 .. code:: json
 
@@ -54,7 +56,7 @@ using the JSON/dict script, it will be :
 
   system = load("system.json")
 
-using RDSystem object constructor :
+Using RDSystem object constructor:
 
 .. code:: python
 
@@ -71,14 +73,15 @@ using RDSystem object constructor :
     state = UnitArray([5, 3], "µmol")
     )
 
-doing it after loading the system :
+Doing it after loading the system:
 
 .. code:: python
 
   system = load_rdsystem("system.json")
   system.state = UnitArray([5, 3], "µmol")
 
-doing it after loading the system, using the set_state method :
+Doing it after loading the system, using the *set_state* method,
+which allow to set the quantity of a given species at a given position:
 
 .. code:: python
 
@@ -144,22 +147,13 @@ doing it after loading the system, using the set_chemostat method :
   system.state.set_chemostat("A", position=(0,0,0), value=5)
   system.state.set_chemoatst("B", position=(0,0,0), value=3)
 
-changing the species density of a system
+Changing the species density of a system
 ----------------------------------------
 
 let us keep the previous example.
-We loaded the system as it was, but now want to change the density of A and B.
-We load the system and set A and B densities and print the system state :
-
-.. code:: python
-
-  system = load_rdsystem("system.json")
-  system.get_species("A").density = 10
-  system.get_species("B").density = 20
-  print(system.state)
-
-however, it will print : [1., 1.] molecule
-For the change to be taken into account, one must regenerate the default system state :
+The system is loaded, but now we want to change the density of the species A and B.
+Just changing A and B's density attributes won't change the system state.
+For the change in density to be taken into account, one must also update the default system state:
 
 .. code:: python
 
@@ -169,4 +163,4 @@ For the change to be taken into account, one must regenerate the default system 
   system.set_default_state()
   print(system.state)
 
-it will now print : [10., 20.] molecule
+it should print : [10., 20.] molecule
