### 2024-03-17 39bdd07
Update setting_up_initial_conditions.rst
