### 2024-03-17 e81fdc3
Update setting_up_initial_conditions.rst
