### 2024-03-17 8e6c191
Update setting_up_initial_conditions.rst
