### 2024-03-17 feb6c93
Update setting_up_initial_conditions.rst
