### 2024-03-17 bad14f8
Update setting_up_initial_conditions.rst
