### 2024-03-17 6709aad
Update setting_up_initial_conditions.rst
