### 2024-03-17 7f13129
Update setting_up_initial_conditions.rst
