### 2024-03-17 a0b4443
Update setting_up_initial_conditions.rst
