### 2024-03-17 f3418aa
Update setting_up_initial_conditions.rst
