### 2024-03-17 13fef8c
Update working_with_output.rst
