### 2024-03-17 1e780d7
Update working_with_output.rst
