### 2024-03-17 e89833c
Update working_with_output.rst
