+++ documentation/using_quantities_with_units.rst
@@ -1,31 +1,34 @@
 working with units
 ==================
 
-UnitsValues and Units
----------------------
+*UnitsValue*, *UnitArray* and *Units*
+-------------------------------------
 
-1) presentation
+1) Presentation
 
-Strengths represents physical quantities with the UnitValue and UnitArray classes, associating respecitvely a numerical
-value or an array of numerical values to some units. Units themselves are represented by the Units class.
+Strengths represents physical quantities with the *UnitValue* and *UnitArray* classes, associating respecitvely a numerical
+value or an array of numerical values to some physical units. Units themselves are represented by the *Units* class.
+Here is an example of how a *UnitValue* obsects can be declared:
 
 .. code:: python
 
   a = UnitValue(5, "µM")
+  a = UnitValue(5, "uM")
   a = UnitValue("5 µM")
   a = parse_unitvalue(5, "µM")
   a = UnitValue(5, Units("µM"))
 
-all the declaration above are equivalent, setting the variable a as 5 micromolar.
+All the declaration above are equivalent, setting the variable a as 5 micromolar.
+Now, examples for the *UnitArray* class:
 
 .. code:: python
 
   a = UnitArray([1,2,3], "µM")
   a = UnitArray([1,2,3], Units("µM"))
 
-while both of the above declarations creates an array of values in micromolar (1 µM, 2 µM and 3 µM).
+Both of the above declarations creates an array of values in micromolar (1 µM, 2 µM and 3 µM).
 
-UnitValue and Units objects can be created from a string, as shown above. the UnitValue string representation
+*UnitValue* and *Units* objects can be created from a string, as shown above. the UnitValue string representation
 is made of a number litteral followed by a Units string representation, separated by one or more whitespaces
 
 .. code:: python
@@ -40,9 +43,8 @@ is made of a number litteral followed by a Units string representation, separate
   UnitValue("[1, 2] µm/s")   # wrong.
   UnitValue("{'v', 1} µm/s") # wrong.
 
-The Units string representation syntax is simple :
-
-units should be a list of symbols separated by "." (multiplication) or "/" (inverse multiplication).
+The syntax for *Units* strings is quite simple:
+Those should be a list of symbols separated by "." (multiplication) or "/" (division).
 
 .. code:: python
 
@@ -50,7 +52,7 @@ units should be a list of symbols separated by "." (multiplication) or "/" (inve
   parse_units("mol/µm. s") # wrong, white space in the unit expression.
   parse_units("mol//µm.s") # wrong, successive "/" have no meaning.
 
-a symbol can be immediately followed by a positive non signed or negative exponent.
+A symbol can be immediately followed by a positive non signed or negative exponent.
 If absent, the exponent is assumed to be 1. The exponent must be an integer.
 
 .. code:: python
@@ -63,7 +65,7 @@ If absent, the exponent is assumed to be 1. The exponent must be an integer.
   parse_units("mol.µm+1.s-2")   # wrong, positive exponents must not be signed
   parse_units("mol.µm 1.s-2")   # wrong, white space in the unit expression.
 
-accepted units symbols are :
+Supported unit symbols are:
 
 +-----+----+--------+------+-------+
 |space|time|quantity|volume|density|
@@ -97,10 +99,10 @@ accepted units symbols are :
 |     |    |molecule|      |       |
 +-----+----+--------+------+-------+
 
-as a consequence, units are case sensitive, since m is meter
-while M is molar (mol/L).
+As a consequence, units are case sensitive, since m corresopnds to meters
+while M corresponds to molars (mol/L).
 
-also, since the µ letter may be inconvenient to type
+Also, since the µ letter may be inconvenient to type
[…]
