### 2024-03-17 4c5f425
Update dependencies.rst
