+++ documentation/dependencies.rst
@@ -6,7 +6,7 @@ Strengths relies on a few external packages.
 NumPy
 -----
 
-A lot of Strengths's features and classes, such as the UnitArray class, rely on the `NumPy <https://numpy.org>`_ package and especially its array type `ndarray <https://numpy.org/doc/stable/reference/arrays.ndarray.html>`_. As a consequence, Strengths requires NumPy to be installed. You can refer to `NumPy's website <https://numpy.org/install/>`_ for more information on how to install NumPy.
+A lot of Strengths's features and classes, such as the UnitArray class, relies on the `NumPy <https://numpy.org>`_ package and especially its array type `ndarray <https://numpy.org/doc/stable/reference/arrays.ndarray.html>`_. As a consequence, Strengths requires NumPy to be installed. You can refer to `NumPy's website <https://numpy.org/install/>`_ for more information on how to install NumPy.
 
 Matplotlib
 ----------
