### 2024-03-17 37d63b8
Update fundamentals.rst
