### 2024-03-17 d2c1ff1
Update fundamentals.rst
