### 2024-03-17 e82997f
Update fundamentals.rst
