### 2024-03-17 e0a3587
Update fundamentals.rst
