### 2024-03-17 9e40e77
Fix documentation (#9)
* Update apiref.rst

updating direct/indirect imports

* Update installing_and_importing_strengths.rst

making use of hyperlinks

* Update README.rst

adding hyperlinks for NumPy and Matplotlib

* Update installing_and_importing_strengths.rst

adding caps and colons

* Update arrays_and_numbers.rst

* Update dependencies.rst

hyperlinks seems more appropriate in this context

* Update dependencies.rst

* Update engines.rst

* Update strengths_and_scipy.rst

* Update building_and_simulating_rds.rst

* Update building_and_simulating_rds.rst

* Update building_and_simulatin
[…]
