+++ documentation/dependencies.rst
@@ -6,26 +6,12 @@ Strengths relies on a few external packages.
 NumPy
 -----
 
-A lot of Strengths features and classes, such as the UnitArray class, rely on the NumPy package (Harris, 2020) [#numpy]_ and especially its famous numpy.ndarray [#numpy_ndarray]_.
-As a consequence, Strenghts require that you have NumPy installed. You can refer to the NumPy website for more information on NumPy :
-
-https://numpy.org
+A lot of Strengths's features and classes, such as the UnitArray class, relies on the `NumPy <https://numpy.org>`_ package and especially its array type `ndarray <https://numpy.org/doc/stable/reference/arrays.ndarray.html>`_. As a consequence, Strengths requires NumPy to be installed. You can refer to `NumPy's website <https://numpy.org/install/>`_ for more information on how to install NumPy.
 
 Matplotlib
 ----------
 
 Strengths provides some functions facilitating the visualization of simulation data with the strengths.plot submodule.
-Those functions rely on the Matplotlib package (Hunter, 2007) [#matplotlib]_, which provides valuable tools for plotting.
+Those functions rely on the `Matplotlib <https://matplotlib.org>`_ package, which provides valuable tools for plotting.
 If you want to use the strengths.plot utility, you'll need to have Matplotlib installed.
-You can refer to the Matplotlib website for more information on Matplotlib :
-
-https://matplotlib.org
-
-References
-----------
-
-.. [#numpy_ndarray] Numpy Developers. numpy 1.25 API reference : numpy.ndarray. (consulted on september 04, 2023). https://numpy.org/doc/stable/reference/generated/numpy.ndarray.html
-
-.. [#matplotlib] Hunter, J. D. (2007). Matplotlib: A 2D graphics environment. Computing in Science \& Engineering, 9(3), 90-95. https://10.1109/MCSE.2007.55
-
-.. [#numpy] Harris, C. R., Millman, K. J., van der Walt, S. J., Gommers, R., Virtanen, P., Cournapeau, D., Wieser, E., Taylor, J., Berg, S., Smith, N. J., Kern, R., Picus, M., Hoyer, S., van Kerkwijk, M. H., Brett, M., Haldane, A., del R{\'{i}}o, J. F., Wiebe, M., Peterson, P., ... Oliphant, T. E. (2020). Array programming with NumPy. Nature, 585(7825), 357-362. https://doi.org/10.1038/s41586-020-2649-2
+You can refer to `Matplotlib's website <https://matplotlib.org/stable/users/installing/index.html>`_ for more information on Matplotlib.
