+++ documentation/simulation.rst
@@ -2,3 +2,4 @@ Reaction Diffusion Simulations
 ==============================
 
 .. autofunction:: strengths.simulate
+.. autofunction:: strengths.simulate_script
