+++ documentation/fundamentals.rst
@@ -1,26 +1,24 @@
 The fundamental concepts
 ========================
 
-object concepts
+Object concepts
 ---------------
 
 Species (Species class)
 ^^^^^^^^^^^^^^^^^^^^^^^
 
-A chemical species is a type of molecule. A system typically comport multiple instances of any species across its space.
+A chemical species is a type of molecule. A system typically contains multiple instances of any species across its space.
 Species can diffuse and be consumed or produced in reactions.
 
 Reaction (Reaction class)
 ^^^^^^^^^^^^^^^^^^^^^^^^^
 
-A chemical transformation, or reaction, is an event localized in space during which some substrate species molecules are converted to some product species molecules.
-Reactions are thus represented through their stoechiometric equation :
+A chemical transformation, or reaction, is an event localized in space during which some substrate species molecules are converted to some product species molecules. Reactions are thus represented by their stoechiometric equation :
 
 .. image:: ABC.png
   :align: center
 
-means that when one molecule of species A collide with a molecule of species B, they can form the a complex, which is another molecule of species C (second order reaction), with a
-reaction rate constant k+, and that the molecule of C can at any moment break down into a molecule of A and a molecule of C, which is the reverse reaction, (first order), whit a rate constant k-.
+In the example above, this means that when one molecule of species A collide with one molecule of species B, they can form the a complex, which is another molecule of species C (second order reaction), with a reaction rate constant :math:`k_+`, and that the molecule of C can at any moment break down into a molecule of A and a molecule of C, which is the reverse reaction, (first order), whit a rate constant :math:`k_-`.
 
 Network (RDNetwork class)
 ^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -53,34 +51,38 @@ Engine (RDEngineBase derived class)
 Reaction diffusion simulation engines are objects that actually carry out the simulations.
 An engine implements some reaction-diffusion algorithm/method.
 
-Output (RDOutput class)
-^^^^^^^^^^^^^^^^^^^^^^^
+Trajectory (RDTrajectory class)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-A reaction diffusion simulation output is the output return by a simulation engine, once the simulation is complete.
-It contains the simulation script that have been used for the simulation as well as the trajectory of its system.
+A reaction diffusion simulation trajectory is the output returned by a simulation engine once the simulation is complete.
+It contains the successive states sampled during the simulations, as well as coontewtual informations such a copy of
+the corresponding system or the simulation script.
 
-other concepts
+Other concepts
 --------------
 
-cells
+Cells
 ^^^^^^
 
 The cell is the unit component of the system space.
-its an individual cubic volume of the system space in which species
+It is a volume of the system space in which species
 can transform and diffuse from/to neighbor cells.
-in a 3D space, a cell have a maximum of 6 neighbors.
+In a 3D space, a cell can be seen as a cube with a maximum of 6 neighbors.
+For the simulation methods currently implemented, species are considered 
+to be distributed homogeneously inside a given cell.
 
-environments
+Environments
 ^^^^^^^^^^^^
 
-an environment is a type of cell.
-some object porperties, such as species initial densities, chemostate and diffusion coefficient,
+An environment is a type of cell.
+Some object porperties, such as species initial densities, chemostate and diffusion coefficient,
 and reaction possibility, can be set to different values depending on the environment.
 This is the feature that enable the creation of inhomogeneous systems.
 
-chemostats
+Chemostats
 ^^^^^^^^^^
 
-a chemostat refers to a process maintaining a species quantity at a constant level.
-any species can be defined as chemostated, globally, in certain environments, or locally (cell wise).
[…]
