### 2024-03-17 ae92706
Update building_and_simulating_rds.rst
fixing example code
