### 2024-03-17 cc542f1
Update rdscript.py
fix
