### 2024-03-17 85d3c17
Thibault fillion patch 1 (#10)
* Update building_and_simulating_rds.rst

fixing example code

* Update rdscript.py

fix

* Update building_and_simulating_rds.rst

* Update building_and_simulating_rds.rst

* Update building_and_simulating_rds.rst

* Update building_and_simulating_rds.rst

* Update building_and_simulating_rds.rst
