### 2024-03-17 6852ce4
Update plot.rst
