+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "strengths"
-version = "0.0.12"
+version = "0.0.13"
 authors = [
 	{name = "Thibault Fillion"},
 	{name = "Francesco Piazza"}]
