### 2024-03-18 ThibaultFillion/strengths#6 @parikshitbajpai
Thanks for adding the instructions. I noticed that the instructions for `matplotlib` have `numpy` in them.

Alternatively, I would suggest keeping a single line of text that says install requirements using `pip install -r requirements.txt`. This way you can ensure that all the requirements are installed in a single step and makes managing requirements easier as the project grows and dependencies increase.
