### 2024-03-18 511a96d
Update MANIFEST.in
