### 2024-03-18 887efbb
Create diffusion_in_graph_spaces.rst
