### 2024-03-18 3abb6f7
Update using_strengths.rst
