### 2024-03-18 db50fa5
Update rdnetwork.py
