+++ pyproject.toml
@@ -1,12 +1,11 @@
 [project]
 name = "strengths"
-version = "0.0.13"
+version = "0.0.14"
 authors = [
 	{name = "Thibault Fillion"},
 	{name = "Francesco Piazza"}]
 description = "A package for the modeling and simulation of reaction-diffusion systems."
 readme = "README.rst"
 dependencies = [
-    "numpy",
-    "matplotlib",
-]
+	"numpy",
+	"matplotlib"]
