### 2024-03-18 ThibaultFillion/strengths#6 @ThibaultFillion
No problem. Indeed, it seems matplotlib requires numpy.
Thank you for your suggestion, this will indeed make the readme and the documentation easier to maintain,
I will make the change now.
