### 2024-03-18 b7c63d3
Update diffusion_in_graph_spaces.rst
