### 2024-03-18 3272ef1
Update diffusion_in_graph_spaces.rst
