### 2024-03-24 f1a566a
Update rdnetwork.py
Fixing an error in the Species class's docstring
