### 2024-03-25 89a5859
Update rdsystem.py
fixing a docstring
