### 2024-03-25 22e5662
Update rdsystem.py
Adding two methods to RDSystem :
* get_state
* get_chemostat
