### 2024-03-25 75d8e35
Update rdsystem.py
fixing the two previously added methods
