### 2024-03-25 8fcb4bc
Update rdsystem.py
adding the apply_reaction method to RDSystem
