### 2024-03-25 efec45b
Update test_rdsystem.py
Adding tests for the new RDSystem method apply_reaction
