+++ tests/test_rdsystem.py
@@ -11,7 +11,7 @@ class SomeCoordClass :
 
 def make_test_rds_1() :
     d_rds = {"network" : {
-        "species" : 
+        "species" :
             [
             {"label" : "A", "density" : "10.5 molecule/µm3", "chstt" : {
                  "cytoplasm" : True,
@@ -28,11 +28,11 @@ def make_test_rds_1() :
                     "outside" : 0.0}
                 }
             ],
-        "reactions" : 
+        "reactions" :
             [
             {"label" : "r", "stoechiometry" : "A -> ", "k+" : "1 s-1", "k-" : 1}
             ],
-        "environments" : 
+        "environments" :
             [
             "membrane",
             "cytoplasm",
@@ -50,7 +50,7 @@ def make_test_rds_1() :
 
 def make_test_rds_2() :
     d_rds = {"network" : {
-        "species" : 
+        "species" :
             [
             {"label" : "A", "density" : "10.5 molecule/µm3", "chstt" : {
                  "cytoplasm" : True,
@@ -67,11 +67,11 @@ def make_test_rds_2() :
                     "outside" : 0.0}
                 }
             ],
-        "reactions" : 
+        "reactions" :
             [
             {"label" : "r", "stoechiometry" : "A -> ", "k+" : "1 s-1", "k-" : 1}
             ],
-        "environments" : 
+        "environments" :
             [
             "membrane",
             "cytoplasm",
@@ -101,7 +101,7 @@ def test_generate_space_cell_env() :
 def test_rds_space_cell_env_set_from_dict() :
     rds = make_test_rds_2()
     assert list(rds.space.cell_env) == [0,0,1,1,2,2]
-    
+
 def test_generate_species_state_default() :
     rds = make_test_rds_1()
     cell_state = generate_species_state(rds.network.get_species("A"), rds.network, rds.space, rds.units_system)
@@ -136,11 +136,11 @@ def test_default_state() :
     assert list(rds.state.value) == [0.0, 0.0,
                                      0.0, 0.0,
                                      0.0, 0.0,
-                                     
+
                                      0.0, 0.0,
                                      0.0, 0.0,
                                      0.0, 0.0,
-                                                                      
+
                                      0.0, 0.0,
                                      0.0, 0.0,
                                      0.0, 0.0,
@@ -158,11 +158,11 @@ def test_default_state() :
     assert list(rds.state.value) == [3.0, 3.0,
                                      3.0, 3.0,
                                      3.0, 3.0,
-                                     
+
                                      3.0, 3.0,
                                      3.0, 3.0,
                                      3.0, 3.0,
-                                                                      
+
                                      3.0, 3.0,
                                      3.0, 3.0,
                                      3.0, 3.0,
@@ -180,11 +180,7 @@ def test_default_state() :
     assert list(rds.state.value) == [3.0, 3.0,
                                      3.0, 3.0,
                                      3.0, 3.0,
-                                     
-                                     3.0, 3.0,
-                                     3.0, 3.0,
-                                     3.0, 3.0,
-                                                                      
+
                                      3.0, 3.0,
                                      3.0, 3.0,
                                      3.0, 3.0,
@@ -192,15 +188,19 @@ def test_default_state() :
                                      3.0, 3.0,
                                      3.0, 3.0,
                                      3.0, 3.0,
-                                     
+
+                                     3.0, 3.0,
+                                     3.0, 3.0,
[…]
