### 2024-03-25 c0aaa9e
Update pyproject.toml
Increment version
