### 2024-04-02 4984835
Update README.rst
* Detailing authors 
* Adding affiliations
