### 2024-04-02 ad365ff
Update README.rst
fix typo
