### 2024-04-05 9f8ba63
Update plot.py
Adding a (keyword only) ax argument to all plot functions
