+++ src/strengths/plot.py
@@ -9,7 +9,7 @@ import math
 plotting utility
 """
 
-def plot_trajectory (output, species, position=None) :
+def plot_trajectory (output, species, position=None, *, ax=None) :
     """
     Plots the time trajectory of one or more species. 
     It relies on the Matplotlib package.
@@ -22,29 +22,37 @@ def plot_trajectory (output, species, position=None) :
     :param position: position of the cell from which we want the trajectory. 
         if None, the global trajectory for the whole system is plotted instead.
     :type position: None, number, tuple or Coord like
+    :param ax: (optional keyword argument) Matplotlib axis to be used (default = None).
     :type species: str or array of str.
     """
+
+    show = False
+    if ax is None :
+        ax = plt.gca()
+        show = True
     
     merge = isnone(position)
     if merge : 
         position=0
     
     if type(species) == str :
-        plt.title(species + " trajectory")
-        plt.ylabel("species quantity (" + str(output.data.units) + ")")
-        plt.xlabel("time (" + str(output.t.units) + ")")
-        plt.plot(output.t.value, output.get_trajectory(species, merge=merge, position=position).value)
-        plt.show()
+        ax.set_title(species + " trajectory")
+        ax.set_ylabel("species quantity (" + str(output.data.units) + ")")
+        ax.set_xlabel("time (" + str(output.t.units) + ")")
+        ax.plot(output.t.value, output.get_trajectory(species, merge=merge, position=position).value)
     else :
-        plt.title("species trajectories")
-        plt.ylabel("species quantity (" + str(output.data.units) + ")")
-        plt.xlabel("time (" + str(output.t.units) + ")")
+        ax.set_title("species trajectories")
+        ax.set_ylabel("species quantity (" + str(output.data.units) + ")")
+        ax.set_xlabel("time (" + str(output.t.units) + ")")
         for s in species: 
-            plt.plot(output.t.value, output.get_trajectory(s, merge=merge, position=position).value, label=s)
-        plt.legend(loc="best")
+            ax.plot(output.t.value, output.get_trajectory(s, merge=merge, position=position).value, label=s)
+        ax.legend(loc="best")
+    if show :
         plt.show()
     
-def plot_sample_state_2D(output, species, sample, axis="auto", axis_position=0, environments=None, xmin=None, xmax=None, units_system=None) : 
+def plot_sample_state_2D(output, species, sample, axis="auto", axis_position=0, 
+                         environments=None, xmin=None, xmax=None, units_system=None,
+                         *, ax=None) : 
     """
     Plots the distribution for the quantity of a species on a plan of the system space.
     It relies on the Matplotlib package.
@@ -66,8 +74,14 @@ def plot_sample_state_2D(output, species, sample, axis="auto", axis_position=0, 
     :type axis: str
     :param axis_position: position of the slice along the slice axis
     :type axis_position: int
+    :param ax: (optional keyword argument) Matplotlib axis to be used (default = None).
     """
-    
+
+    show = False
+    if ax is None :
+        ax = plt.gca()
+        show = True
+        
     if type(output.system.space) != RDGridSpace :
         raise ValueError("plot_sample_state_2D requires the system to have a grid space (RDGridSpace).")
     
[…]
