### 2024-04-14 d524d40
Fix typos in paper for JOSS review (#6495)
