### 2024-04-16 #6495 @parikshitbajpai
Hello @ThibaultFillion,

I am almost done with the review but have a few doubts I hope you can clarify.

1. Are the other similar tools specific to certain applications? I feel there are other general purpose PDE solvers which could be used for modelling reaction-diffusion system, e.g. FEM codes. Could you please clarify the criteria you used.
2. Are the examples meant to be original results or are they demonstrating the capability? I might have missed it in the manuscript if they were meant to be original results, and, if not, can you please clarify if the examples are based on some previously available models / results since that was not referenced?
3. Would you consider adding the examples scripts to the repo? It will enable users and people reading the article to quickly be able to verify the showcased capabilities.

Thanks!
