### 2024-04-17 #6495 @jakryd
The limitations of the package are explained in the paper, but I think that it can still be useful, especially for students who are starting to learn about such algorithms. The examples shown in @ThibaultFillion's paper are particularly compelling, and it would be great to make them available as Jupyter Notebooks so that users can experiment with them. But this can be done after publication. Overall, I did not come across any issues with the current version, and I am happy to recommend publication without any further modifications.
