### 2024-04-18 ae20ae8
Typos etc. spotted during JOSS paper review  (#13)
* Fix typos in paper for JOSS review (#6495)

* More typo catches / edit suggestions for JOSS review (#6495)
