+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title : STReNGTHS, a python package to model and simulate reaction-diffusion systems
+title : STReNGTHS, a Python package to model and simulate reaction-diffusion systems
 authors :
   - name : Thibault Fillion
     affiliation : "1, 2"
@@ -19,11 +19,11 @@ bibliography : references.bib
 # Summary
 
 STReNGTHS is an open-source Python package that provides a simple and intuitive
-interface for designing models of discrete 3D heterogeneous reaction-diffusion systems and simulating their trajectories. Different algorithms are available, both stochastic (exact or approximate  solutions of the associated 
-master equation) or deterministic (numerical solutions of the corresponding rate equations). 
-The acronym stands for "Simulation and modeling Tool for REaction-diffusion Networks in Graphs and Tridimensional Heterogeneous Systems" (STReNGTHS). The simulation algorithms are interfaced through a general abstract interface, which makes it easy to extend STReNGTHS with new algorithms and other features.  It is implemented in python (standard library, 
-Numpy [@harris_numpy_2020] and Matplotlib [@hunter_matplotlib_2007]) 
-and C++ (standard C++11 or later), and can be easily installed from the Python Package Index (PyPI, https://pypi.org) with (i.e.)
+interface for designing models of discrete 3D heterogeneous reaction-diffusion systems and simulating their trajectories. Different algorithms are available, both stochastic (exact or approximate  solutions of the associated
+master equation) and deterministic (numerical solutions of the corresponding rate equations).
+The acronym stands for "Simulation and modeling Tool for REaction-diffusion Networks in Graphs and Tridimensional Heterogeneous Systems" (STReNGTHS). The simulation algorithms are interfaced through a general abstract interface, which makes it easy to extend STReNGTHS with new algorithms and other features.  It is implemented in Python (standard library,
+Numpy [@harris_numpy_2020] and Matplotlib [@hunter_matplotlib_2007])
+and C++ (standard C++11 or later), and can be easily installed from the Python Package Index (PyPI, <https://pypi.org>) with (i.e.)
 
 \begin{verbatim}
     pip install strengths
@@ -31,171 +31,171 @@ and C++ (standard C++11 or later), and can be easily installed from the Python P
 
 # Statement of need
 
-Biology at the cell scale relies on complex networks of biochemical reactions, 
+Biology at the cell scale relies on complex networks of biochemical reactions,
 usually operating across multiple membrane-associated as well as
 membrane-less compartments (i.e. plasma membrane, cytoplasm, cellular
-organelles, stress granules, etc) and driven far from equilibrium by highly regulated species, 
-such as nucleotides (ATP/GTP), amino acids and different 
-ions. In order to understand the properties of such reaction-diffusion 
-networks, and especially their role in macroscopic *emergent* 
-phenomena, convenient, reliable and efficient modeling and simulation tools supporting heterogeneous systems are necessary. Moreover, the choice of the simulation algorithm to be used may depend on the system : deterministic approaches such as ODE integration (rate equations) are effective, but inappropriate for systems that are sensitive to fluctuations, 
-such as systems that operate with species present at low copy numbers 
+organelles, stress granules, etc) and driven far from equilibrium by highly regulated species,
+such as nucleotides (ATP/GTP), amino acids and different
+ions. In order to understand the properties of such reaction-diffusion
+networks, and especially their role in macroscopic *emergent*
[…]
