### 2024-04-19 b232e21
Update paper.md
Adding references for the diffusion coefficients in example 1
