### 2024-04-19 feb85db
Update references.bib
Adding references
Homogeneizing indentation
