+++ paper/references.bib
@@ -13,7 +13,7 @@
 	note = {Publisher: American Chemical Society},
 	pages = {2340--2361},
 	file = {ACS Full Text Snapshot:C\:\\Users\\33782\\Zotero\\storage\\V4DHLM9P\\j100540a008.html:text/html},
-}
+	}
 
 @article{bernstein_simulating_2005,
 	title = {Simulating mesoscopic reaction-diffusion systems using the {Gillespie} algorithm},
@@ -26,7 +26,7 @@
 	year = {2005},
 	pages = {041103},
 	file = {Full Text PDF:C\:\\Users\\33782\\Zotero\\storage\\6QZ2PUK4\\Bernstein - 2005 - Simulating mesoscopic reaction-diffusion systems u.pdf:application/pdf},
-}
+	}
 
 @article{gillespie_approximate_2001,
 	title = {Approximate {Accelerated} {Stochastic} {Simulation} of {Chemically} {Reacting} {Systems}},
@@ -39,85 +39,75 @@
 	year = {2001},
 	pages = {1716--1733},
 	file = {Version soumise:C\:\\Users\\33782\\Zotero\\storage\\9EJK7RDW\\Gillespie - 2001 - Approximate Accelerated Stochastic Simulation of C.pdf:application/pdf},
-}
+	}
 
 @article{hoffmann2019readdy,
-  title={ReaDDy 2: Fast and flexible software framework for interacting-particle reaction dynamics},
-  author={Hoffmann, Moritz and Fr{\"o}hner, Christoph and No{\'e}, Frank},
-  journal={PLoS computational biology},
-  volume={15},
-  number={2},
-  pages={e1006830},
-  year={2019},
-  publisher={Public Library of Science},
-  doi={10.1371/journal.pcbi.1006830}
-}
+	title={ReaDDy 2: Fast and flexible software framework for interacting-particle reaction dynamics},
+	author={Hoffmann, Moritz and Fr{\"o}hner, Christoph and No{\'e}, Frank},
+	journal={PLoS computational biology},
+	volume={15},
+	number={2},
+	pages={e1006830},
+	year={2019},
+	publisher={Public Library of Science},
+	doi={10.1371/journal.pcbi.1006830}
+	}
 
 @article{10.1093/bioinformatics/bti431,
-    author = {Hattne, Johan and Fange, David and Elf, Johan},
-    title = "{Stochastic reaction-diffusion simulation with MesoRD}",
-    journal = {Bioinformatics},
-    volume = {21},
-    number = {12},
-    pages = {2923-2924},
-    year = {2005},
-    month = {04},
-    abstract = "{Summary: MesoRD is a tool for stochastic simulation of chemical reactions and diffusion. In particular, it is an implementation of the next subvolume method, which is an exact method to simulate the Markov process corresponding to the reaction-diffusion master equation.Availability: MesoRD is free software, written in C++ and licensed under the GNU general public license (GPL). MesoRD runs on Linux, Mac OS X, NetBSD, Solaris and Windows XP. It can be downloaded from http://mesord.sourceforge.net.Contact:  johan.elf@icm.uu.se; johan.hattne@embl-hamburg.deSupplementary information: ‘MesoRD User's Guide’ and other documents are available at http://mesord.sourceforge.net.}",
-    issn = {1367-4803},
-    doi = {10.1093/bioinformatics/bti431},
-    url = {https://doi.org/10.1093/bioinformatics/bti431},
-    eprint = {https://academic.oup.com/bioinformatics/article-pdf/21/12/2923/48971060/bioinformatics\_21\_12\_2923.pdf},
-}
+	author = {Hattne, Johan and Fange, David and Elf, Johan},
[…]
