+++ paper/references.bib
@@ -145,7 +145,7 @@
 	}
 
 @book{Milo2015,
-	title={CELL BIOLOGY by the numbers},
+	title={Cell biology by the numbers},
 	author={Milo, Ron and Phillips, Rob},
 	pages={258},
 	edition={1},
