+++ paper/examples/example1/README.rst
@@ -1,0 +1,58 @@
+Example 1
+=========
+
+The folder contains the simulation script for the first example of the paper.
+It contains the following files :
+*  README.rst : This file.
+*  cg_map.txt : Coarse graining map utilized for this example.
+*  environments.txt : Layout of the environment associated with each cell of the system
+*  system.json : JSON description of the reaction-diffusion system, loaded by the example.
+*  script.py : Simulation script for the non coarse-grained simulations.
+*  script_cg.py : Simulation script for the coarse-grained simulations.
+*  results.py : Script plotting the output generated by the two simulation scripts.
+as well as output files :
+*  traj_cg_eul2_data.npy
+*  traj_cg_eul2.json
+*  traj_cg_tlp2_data.npy
+*  traj_cg_tlp2.json
+*  traj_eul2_data.npy
+*  traj_eul2.json
+*  traj_tlp2_data.npy
+*  traj_tlp2.json
+
+To run this example, first step is to run script.py and script_cg.py in the directory.
+Those will output JSON and NumPy files inside the folder. Second and last step is to run results.py, which will plot the results using Matplotlib.
+The original results, which are already present inside the folder, may be displayed by directly running results.py.
+
+Note: the original results (JSON and NumPy files) inside the directory were generated with an early version of the package and of the script files,
+and some files were renamed. The instruction
+
+.. code:: python
+
+  "tlpcgucg = strncg.uncoarsegrain_trajectory(tlpcg, tlp.system.space, cgmap)"
+
+(results.py, l.52) was replaced by
+
+.. code:: python
+
+  "tlpcgucg = strncg.uncoarsegrain_trajectory(tlpcg, tlp.system, cgmap)"
+
+due to an API change. Also, doing the coarse-graining by hand as in the script files is not necessary
+anymore, as the coarse-graining map can simply be passed as an argument to the simulate function.
+
+Reaction kinetic parameters are guessed.
+Diffusion parameters are set by order of magnitude from
+references [1] and [2], as explained in the paper :
+*R* and *C* are supposed to be membrane receptors, and their diffusion
+coefficients are thus set in the order of magnitude of that of a transmembrane
+protein diffusing in a compartmented plasma membrane [2].
+*L* may me a small protein, thus its diffusion coefficient is set to that of a
+protein of about 30 kDa diffusing in water [1].
+*X* and *Y* are also proteins, but diffuse inside inside of the cell, thus
+their diffusion coefficients are set to that of a protein of about 30 kDa
+diffusing in cytosol [1].
+
+References
+----------
+* [1] Milo, R., & Phillips, R. (2015). *Cell biology by the numbers* (1st ed.). Garland Science, Taylor & Francis Group. https://doi.org/10.1201/9780429258770
+* [2] Fujiwara, T. K., Iwasawa, K., Kalay, Z., Tsunoyama, T. A., Watanabe, Y., Umemura, Y. M., Murakoshi, H., Suzuki, K. G. N., Nemoto, Y. L., Morone, N., & Kusumi, A. (2016). Confined diffusion of transmembrane proteins and lipids induced by the same actin meshwork lining the plasma membrane. *Molecular Biology of the Cell*, *27*(7), 1101-1119. https://doi.org/10.1091/mbc.E15-04-0186
