+++ paper/examples/example1/cg_map.txt
@@ -1,0 +1,26 @@
+0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,1   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,0   ,16  ,17  ,18  ,19  ,20  ,21  ,22  ,23  ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,0   ,80  ,81  ,82  ,83  ,8   ,8   ,8   ,8   ,8   ,8   ,24  ,25  ,26  ,27  ,2   ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,78  ,79  ,15  ,15  ,15  ,8   ,8   ,8   ,8   ,8   ,8   ,9   ,9   ,9   ,28  ,29  ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,77  ,15  ,15  ,15  ,15  ,15  ,8   ,8   ,8   ,8   ,9   ,9   ,9   ,9   ,9   ,30  ,2   ,2   ,2   ,2   ,2   ,
+0   ,0   ,0   ,0   ,0   ,76  ,15  ,15  ,15  ,15  ,15  ,8   ,8   ,8   ,8   ,9   ,9   ,9   ,9   ,9   ,31  ,2   ,2   ,2   ,2   ,2   ,
+3   ,3   ,3   ,3   ,74  ,75  ,15  ,15  ,15  ,15  ,15  ,84  ,84  ,84  ,84  ,9   ,9   ,9   ,9   ,9   ,32  ,33  ,4   ,4   ,4   ,4   ,
+3   ,3   ,3   ,3   ,73  ,14  ,14  ,15  ,15  ,15  ,84  ,84  ,84  ,84  ,84  ,84  ,9   ,9   ,9   ,10  ,10  ,34  ,4   ,4   ,4   ,4   ,
+3   ,3   ,3   ,3   ,72  ,14  ,14  ,14  ,14  ,84  ,84  ,84  ,84  ,84  ,84  ,84  ,84  ,10  ,10  ,10  ,10  ,35  ,4   ,4   ,4   ,4   ,
+3   ,3   ,3   ,3   ,71  ,14  ,14  ,14  ,14  ,84  ,84  ,84  ,84  ,84  ,84  ,84  ,84  ,10  ,10  ,10  ,10  ,36  ,4   ,4   ,4   ,4   ,
[…]
