### 2024-04-21 afa009b
Update README.rst
removing output files because too large
