+++ paper/examples/example1/README.rst
@@ -10,19 +10,9 @@ It contains the following files :
 *  script.py : Simulation script for the non coarse-grained simulations.
 *  script_cg.py : Simulation script for the coarse-grained simulations.
 *  results.py : Script plotting the output generated by the two simulation scripts.
-as well as output files :
-*  traj_cg_eul2_data.npy
-*  traj_cg_eul2.json
-*  traj_cg_tlp2_data.npy
-*  traj_cg_tlp2.json
-*  traj_eul2_data.npy
-*  traj_eul2.json
-*  traj_tlp2_data.npy
-*  traj_tlp2.json
 
 To run this example, first step is to run script.py and script_cg.py in the directory.
 Those will output JSON and NumPy files inside the folder. Second and last step is to run results.py, which will plot the results using Matplotlib.
-The original results, which are already present inside the folder, may be displayed by directly running results.py.
 
 Note: the original results (JSON and NumPy files) inside the directory were generated with an early version of the package and of the script files,
 and some files were renamed. The instruction
