### 2024-04-21 0a2ad56
Delete paper/examples/example2/README.md
duplicated file
