### 2024-04-21 3599ddc
Update README.rst
removing output files because too large
