### 2024-04-21 0d09cc3
Paper changes (#14)
* Update paper.md

* Update paper.md

* Delete paper/example1.png

replacing example1.png #1

* Add files via upload

#replacing example1.png #2
Adding references for the diffusion coefficients

* Add files via upload

Replacing example1.png #2
Adding references for the diffusion coefficients

* Update paper.md

Adding references for the diffusion coefficients in example 1

* Update references.bib

Adding references
Homogeneizing indentation

* Update references.bib

* Create README.rst

* Create README.md

* Add files via upload

* Add files via upload

* Delete paper/exam
[…]
