### 2024-04-23 42bd73c
Update paper.md
Adding references to the documentation.
