### 2024-04-23 23ca097
Update references.bib
Adding a reference to the documentation
