### 2024-04-23 ff18d7c
Update references.bib
