+++ paper/references.bib
@@ -170,11 +170,11 @@
 	eprint = {https://doi.org/10.1091/mbc.E15-04-0186}
 	}
 
-@misc {Strengths_doc,
+@misc{Strengths_doc,
 	author = {Fillion, Thibault and Piazza, Francesco},
 	title = {Building a reaction diffusion system and performing simulations. Documentation for Strengths},
 	month = Apr,
 	year = 2024,
-	note = {Accessed online on the 23rd of april, 2024.}
-	url = {https://strengths.readthedocs.io/en/latest/building_and_simulating_rds.html},
+	note = {Accessed online on the 23rd of april, 2024},
+	url = {https://strengths.readthedocs.io/en/latest/building_and_simulating_rds.html}
 	}
