+++ paper/references.bib
@@ -173,7 +173,7 @@
 @misc{Strengths_doc,
 	author = {Fillion, Thibault and Piazza, Francesco},
 	title = {Building a reaction diffusion system and performing simulations. Documentation for Strengths},
-	month = Apr,
+	month = apr,
 	year = 2024,
 	note = {Accessed online on the 23rd of april, 2024},
 	url = {https://strengths.readthedocs.io/en/latest/building_and_simulating_rds.html}
