+++ paper/references.bib
@@ -11,34 +11,29 @@
 	month = dec,
 	year = {1977},
 	note = {Publisher: American Chemical Society},
-	pages = {2340--2361},
-	file = {ACS Full Text Snapshot:C\:\\Users\\33782\\Zotero\\storage\\V4DHLM9P\\j100540a008.html:text/html},
+	pages = {2340--2361}
 	}
 
 @article{bernstein_simulating_2005,
 	title = {Simulating mesoscopic reaction-diffusion systems using the {Gillespie} algorithm},
 	volume = {71},
 	doi = {10.1103/PhysRevE.71.041103},
-	abstract = {We examine an application of the Gillespie algorithm to simulating spatially inhomogeneous reaction-diffusion systems in mesoscopic volumes such as cells and microchambers. The method involves discretizing the chamber into elements and modeling the diffusion of chemical species by the movement of molecules between neighboring elements. These transitions are expressed in the form of a set of reactions which are added to the chemical system. The derivation of the rates of these diffusion reactions is by comparison with a finite volume discretization of the heat equation on an unevenly spaced grid. The diffusion coefficient of each species is allowed to be inhomogeneous in space, including discontinuities. The resulting system is solved by the Gillespie algorithm using the fast direct method. We show that in an appropriate limit the method reproduces exact solutions of the heat equation for a purely diffusive system and the nonlinear reaction-rate equation describing the cubic autocatalytic reaction.},
 	journal = {Physical review. E, Statistical, nonlinear, and soft matter physics},
 	author = {Bernstein, David},
 	month = may,
 	year = {2005},
-	pages = {041103},
-	file = {Full Text PDF:C\:\\Users\\33782\\Zotero\\storage\\6QZ2PUK4\\Bernstein - 2005 - Simulating mesoscopic reaction-diffusion systems u.pdf:application/pdf},
+	pages = {041103}
 	}
 
 @article{gillespie_approximate_2001,
 	title = {Approximate {Accelerated} {Stochastic} {Simulation} of {Chemically} {Reacting} {Systems}},
 	volume = {115},
 	doi = {10.1063/1.1378322},
-	abstract = {The stochastic simulation algorithm (SSA) is an essentially exact procedure for numerically simulating the time evolution of a well-stirred chemically reacting system. Despite recent major improvements in the efficiency of the SSA, its drawback remains the great amount of computer time that is often required to simulate a desired amount of system time. Presented here is the “τ-leap” method, an approximate procedure that in some circumstances can produce significant gains in simulation speed with acceptable losses in accuracy. Some primitive strategies for control parameter selection and error mitigation for the τ-leap method are described, and simulation results for two simple model systems are exhibited. With further refinement, the τ-leap method should provide a viable way of segueing from the exact SSA to the approximate chemical Langevin equation, and thence to the conventional deterministic reaction rate equation, as the system size becomes larger.},
 	journal = {Journal of Chemical Physics},
 	author = {Gillespie, Daniel},
 	month = jul,
 	year = {2001},
-	pages = {1716--1733},
-	file = {Version soumise:C\:\\Users\\33782\\Zotero\\storage\\9EJK7RDW\\Gillespie - 2001 - Approximate Accelerated Stochastic Simulation of C.pdf:application/pdf},
+	pages = {1716--1733}
 	}
 
 @article{hoffmann2019readdy,
@@ -62,7 +57,6 @@
 	pages = {2923-2924},
 	year = {2005},
 	month = {04},
[…]
