+++ paper/references.bib
@@ -60,7 +60,7 @@
 	issn = {1367-4803},
 	doi = {10.1093/bioinformatics/bti431},
 	url = {https://doi.org/10.1093/bioinformatics/bti431},
-	eprint = {https://academic.oup.com/bioinformatics/article-pdf/21/12/2923/48971060/bioinformatics\_21\_12\_2923.pdf},
+	eprint = {https://academic.oup.com/bioinformatics/article-pdf/21/12/2923/48971060/bioinformatics\_21\_12\_2923.pdf}
 	}
 
 @article{Elf2004,
@@ -87,7 +87,7 @@
 	issn = {1367-4803},
 	doi = {10.1093/bioinformatics/btw469},
 	url = {https://doi.org/10.1093/bioinformatics/btw469},
-	eprint = {https://academic.oup.com/bioinformatics/article-pdf/32/21/3366/49021661/bioinformatics\_32\_21\_3366.pdf},
+	eprint = {https://academic.oup.com/bioinformatics/article-pdf/32/21/3366/49021661/bioinformatics\_32\_21\_3366.pdf}
 	}
 
 @article{10.3389/neuro.11.015.2009,  
@@ -129,7 +129,7 @@
 	author = {Hunter, J. D.},
 	year = {2007},
 	note = {Publisher: IEEE COMPUTER SOC},
-	pages = {90--95},
+	pages = {90--95}
 	}
 
 @book{Milo2015,
