### 2024-04-23 1467635
Paper changes 2 (#15)
* Update paper.md

Adding references to the documentation.

* Update references.bib

Adding a reference to the documentation

* Update references.bib

* Update references.bib

* Update references.bib

* Update references.bib

* Delete paper/example2.png

Improving fig. 2 (a) #1

* Add files via upload

Improving fig. 2 (a) #2
