### 2024-04-23 #6495 @ThibaultFillion
Thank you for your reviews @parikshitbajpai and @jakryd .

@parikshitbajpai Thank you for raising your concerns.
1) The similar tools evoked in the paper are all tools applied to the modeling and simulation of biochemical systems, providing an abstract interface above the lower level considerations such as ODE or PDE solving.
2) The figures in the paper are original results but are solely here for the purpose of illustrating some of the different features of the package. The first considers a fictive signal transduction scheme.
The reaction rates are reasonable guesses but the diffusion coefficients are set from the literature and represent the correct order of magnitude, based on how and where the different species diffuse.
We have added references for this in the manuscript.
The second model applies some generic pattern-forming network to systems with various shapes.
We added references to the documentation of the package as similar pattern-forming reaction-diffusion networks are used there.
3) Indeed, we added the simulation scripts to the paper repository.

@jakryd Thank you for your review. Making examples available as Jupyter notebooks would be a good idea. For now, we added the simulation scripts for the two examples to the paper repository.
