### 2024-05-08 ead21e6
Update references.bib
adding two references : 
* McGough2004
* Ruijgrok1997
