### 2024-05-08 ea6857b
Update paper.md
Adding references to 
McGough2004 and Ruijgrok1997
in the second example.
