### 2024-05-08 #6495 @ThibaultFillion
Hello @Nikoleta-v3 , @parikshitbajpai and @jakryd

We would like to add two more references to the paper for the second example.
Here is the branch with the changes:
https://github.com/ThibaultFillion/strengths/tree/paper_addrefs
Would you agree for this branch to be merged with the "paper" one ?
