### 2024-05-08 3706c16
Update building_and_simulating_rds.rst
Adding references for the pattern-making reaction network in the examples.
