+++ documentation/building_and_simulating_rds.rst
@@ -720,6 +720,6 @@ Let us see what happens when we apply different boundary conditions to this syst
 References
 ----------
 
-.. [#McGough2004] McGough, Jeff S. & Riley, K. (2004). Pattern formation in the Gray–Scott model.\ *Nonlinear Analysis: Real World Applications*\ , \ *5*\ (1). 105-121. https://doi.org/10.1016/S1468-1218(03)00020-8
+.. [#McGough2004] McGough, Jeff S. & Riley, K. (2004). Pattern formation in the Gray–Scott model. \ *Nonlinear Analysis: Real World Applications*\ , \ *5*\ (1). 105-121. https://doi.org/10.1016/S1468-1218(03)00020-8
 
 .. [#Ruijgrok1997] Ruijgrok, Th & Ruijgrok, M. (1997). A reaction-diffusion equation for a cyclic system with three components. \ *Journal of Statistical Physics*\ , \ *97*\ . 1145-1164. https://doi.org/10.1007/BF02181277
