### 2024-05-08 a0c8963
Update references.bib
updating the documentation reference access date
