+++ paper/references.bib
@@ -161,9 +161,9 @@
 @misc{Strengths_doc,
 	author = {Fillion, Thibault and Piazza, Francesco},
 	title = {Building a reaction diffusion system and performing simulations. Documentation for Strengths},
-	month = apr,
+	month = may,
 	year = 2024,
-	note = {Accessed online on the 23rd of april, 2024},
+	note = {Accessed online on the 8th of may, 2024},
 	url = {https://strengths.readthedocs.io/en/latest/building_and_simulating_rds.html}
 	}
 
