### 2024-05-08 #6495 @parikshitbajpai
> Would you agree for this branch to be merged with the "paper" one ?

Hey @ThibaultFillion, this looks good to me and can be merged as long as @Nikoleta-v3 and @jakryd are okay with it as well.
