### 2024-05-10 #6495 @jakryd
> Hello @Nikoleta-v3 , @parikshitbajpai and @jakryd
>
> We would like to add two more references to the paper for the second example. Here is the branch with the changes: https://github.com/ThibaultFillion/strengths/tree/paper_addrefs Would you agree for this branch to be merged with the "paper" one ?
>
> Those references were also added to the documentation.

Sure!
