### 2024-05-18 #6495 @Nikoleta-v3
Hello @ThibaultFillion,

I am very sorry for the long silence and the delay in processing your submission. Life has been hectic the past few weeks. Regardless, apologies for the delay.

> @Nikoleta-v3 Would you agree with these changes?

Yes, I agree with the changes, thank you.

Regarding the submission, I had some time to go over it as well. The package looks great, very useful, and your documents page is very well implemented.

I believe there is a typo in the paper. In the Figure 2 caption, it should be `stoichiometric` instead of `stoechiometric`. Otherwise, I only have two suggestions (but feel free to not implement them):

1. To run the tests, one has to run the script `run_all_tests.py`, which just calls all the test functions you have implemented. Since all tests are functions, you could potentially use the package `pytest` (https://docs.pytest.org/en/8.2.x/) to run the tests instead. What's nice about `pytest` is that it gives nice output. For example, this is me running the command `pytest` from your tests folder. There are a few failures, but it has to do with the relative paths. If you would like to use `pytest`, I would be happy to help you set things up.
![Screenshot 2024-05-18 at 22 26 35 (2)](https://github.com/openjournals/joss-reviews/assets/19708408/0aa0dd5a-0ff5-4b4a-9c7f-75d1a42b14dd)

2. Regarding the documentation, in the `Using Strengths` section, there are several subsections, but the user has to scroll to see all of them. I was just thinking if there is a way to make it a bit easier for the user to see all the sections without having to scroll. One idea would be to list all the sections at the top of the page.

Again, all these comments are minor.
