### 2024-05-20 0385a73
Paper addrefs (#18)
* Update references.bib

adding two references : 
* McGough2004
* Ruijgrok1997

* Update paper.md

Adding references to 
McGough2004 and Ruijgrok1997
in the second example.

* Update references.bib

updating the documentation reference access date

* Update paper.md

fix
