+++ paper/references.bib
@@ -161,8 +161,32 @@
 @misc{Strengths_doc,
 	author = {Fillion, Thibault and Piazza, Francesco},
 	title = {Building a reaction diffusion system and performing simulations. Documentation for Strengths},
-	month = apr,
+	month = may,
 	year = 2024,
-	note = {Accessed online on the 23rd of april, 2024},
+	note = {Accessed online on the 8th of may, 2024},
 	url = {https://strengths.readthedocs.io/en/latest/building_and_simulating_rds.html}
 	}
+
+@article{McGough2004,
+	title = {Pattern formation in the {Gray}–{Scott} model},
+	volume = {5},
+	issn = {1468-1218},
+	doi = {10.1016/S1468-1218(03)00020-8},
+	number = {1},
+	journal = {Nonlinear Analysis: Real World Applications},
+	author = {McGough, Jeff S. and Riley, Kyle},
+	month = feb,
+	year = {2004},
+	pages = {105--121}
+  	}
+
+@article{Ruijgrok1997,
+	title = {A reaction-diffusion equation for a cyclic system with three components},
+	volume = {87},
+	doi = {10.1007/BF02181277},
+	journal = {Journal of Statistical Physics},
+	author = {Ruijgrok, Th and Ruijgrok, M.},
+	month = jun,
+	year = {1997},
+	pages = {1145--1164}
+  	}
