### 2024-05-20 e2acb29
Paper (#19)
* Create paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Add files via upload

* Add files via upload

* Create references.bib

* Update references.bib

* Update paper.md

* Update paper.md

* Create draft-pdf.yml

* Update draft-pdf.yml

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update draft-pdf.yml

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md


[…]
