+++ paper/paper.md
@@ -1,0 +1,293 @@
+---
+title : STReNGTHS, a Python package to model and simulate reaction-diffusion systems
+authors :
+  - name : Thibault Fillion
+    affiliation : "1, 2"
+  - name : Francesco Piazza
+    affiliation : "3, 4"
+affiliations :
+  - name : Université d'Orléans, UFR Sciences Techniques, Avenue du Parc Floral, BP 6749, 45067 Orléans, France
+    index : 1
+  - name : Centre de Biophysique Moléculaire (CBM), CNRS-UPR 4301, Rue Charles Sadron, 45071 Orléans, France
+    index : 2
+  - name : Dipartimento di fisica & Astronomia, Università di Firenze, Via G. Sansone 1, 50019 Sesto Fiorentino, Italy
+    index : 3
+  - name : INFN sezione di Firenze, Via G. Sansone 1, 50019 Sesto Fiorentino, Italy
+    index : 4
+bibliography : references.bib
+---
+# Summary
+
+STReNGTHS is an open-source Python package that provides a simple and intuitive
+interface for designing models of discrete 3D heterogeneous reaction-diffusion systems and simulating their trajectories. Different algorithms are available, both stochastic (exact or approximate  solutions of the associated
+master equation) and deterministic (numerical solutions of the corresponding rate equations).
+The acronym stands for "Simulation and modeling Tool for REaction-diffusion Networks in Graphs and Tridimensional Heterogeneous Systems" (STReNGTHS). The simulation algorithms are interfaced through a general abstract interface, which makes it easy to extend STReNGTHS with new algorithms and other features.  It is implemented in Python (standard library,
+Numpy [@harris_numpy_2020] and Matplotlib [@hunter_matplotlib_2007])
+and C++ (standard C++11 or later), and can be easily installed from the Python Package Index (PyPI, <https://pypi.org>) with (i.e.)
+
+\begin{verbatim}
+    pip install strengths
+\end{verbatim}
+
+# Statement of need
+
+Biology at the cell scale relies on complex networks of biochemical reactions,
+usually operating across multiple membrane-associated as well as
+membrane-less compartments (i.e. plasma membrane, cytoplasm, cellular
+organelles, stress granules, etc) and driven far from equilibrium by highly regulated species,
+such as nucleotides (ATP/GTP), amino acids and different
+ions. In order to understand the properties of such reaction-diffusion
+networks, and especially their role in macroscopic *emergent*
+phenomena, convenient, reliable and efficient modeling and simulation tools supporting heterogeneous systems are necessary. Moreover, the choice of the simulation algorithm to be used may depend on the system : deterministic approaches such as ODE integration (rate equations) are effective, but inappropriate for systems that are sensitive to fluctuations,
+such as systems that operate with species present at low copy numbers
+(e.g. certain enzymes, many mRNA species) and/or
+in tiny reaction volumes (i.e. within mitochondria),
+for which stochastic methods should be preferred.
+This is why one needs to have multiple, flexible approaches available.
+STReNGTHS provides an interface to simulate reaction-diffusion systems
+and manipulate their trajectories, as well as full control
+and access to the simulation algorithms themselves.
+
+Reaction-diffusion systems consist primarily
+of a reaction-diffusion network, which is represented by
+the *RDSystem* class. This defines a set of coupled chemical transformations
+that is supplemented with a spatial distribution of
+chemical species, whose features describe
+where and how fast each molecule shall diffuse.
+The system space is discrete, and consists of a 3D mesh of individual
+volume elements, which we refer to as *cells*.
[…]
