### 2024-05-20 f1b8d21
Update test_kinetics.py
