+++ tests/test_kinetics.py
@@ -11,9 +11,9 @@ import numpy as np
 # test reaction rate
 # should be done both for rdnetworks and graphs
 
-def test_compute_reaction_rates_dimensions_for_different_reaction_orders_with_default_units_default_units(spacetype) :
+def _test_compute_reaction_rates_dimensions_for_different_reaction_orders_with_default_units_default_units(spacetype) :
     #checks that rates are in molecule/s, no mater the reaction order
- 
+
     system_dict = {
         "network" : {
             "species" : [
@@ -55,7 +55,7 @@ def test_compute_reaction_rates_dimensions_for_different_reaction_orders_with_de
         for rate in rates :
             assert rate.units.dim == UnitsDimensions(quantity=1, space=0, time=-1)
 
-def test_compute_reaction_rates_value(spacetype) :
+def _test_compute_reaction_rates_value(spacetype) :
     #checks that rates computed values are correct
     kf = 12
     kr = 0.5
@@ -67,9 +67,9 @@ def test_compute_reaction_rates_value(spacetype) :
     D = 548
     E = 0.33
     F = 1000
-    
+
     NA = avogadro_number()
-    state = UnitArray([A, B, C, D, E, F], "mol")    
+    state = UnitArray([A, B, C, D, E, F], "mol")
 
     system_dict = {
         "network" : {
@@ -94,13 +94,13 @@ def test_compute_reaction_rates_value(spacetype) :
             "edges" : []}
     else :
         system_dict["space"] = {"cell_vol" : str(V)+" mL"}
-        
+
     system = rdsystem_from_dict(system_dict)
 
     V *= 1e12
     rf = kf/60 * V * NA*A/V * (NA*C/V)**2 * (NA*B/V)**4 # molecule/s
     rr = kr * V * (NA*D/V)**6 * (NA*F/V)**3             # molecule/s
-    
+
     rates = strnkin.compute_reaction_rates(
         system = system,
         reaction = system.network.reactions[0],
@@ -111,10 +111,10 @@ def test_compute_reaction_rates_value(spacetype) :
     assert np.isclose(rates[0].convert("molecule/s").value, rf)
     assert np.isclose(rates[1].convert("molecule/s").value, rr)
 
-def test_compute_reaction_rates_with_environments(spacetype) :
+def _test_compute_reaction_rates_with_environments(spacetype) :
     #checks that reaction env are taken into account
 
-    state = UnitArray([1, 1, 1, 1], "molecule")        
+    state = UnitArray([1, 1, 1, 1], "molecule")
     system_dict = {
         "network" : {
             "environments" : ["a", "b"],
@@ -137,7 +137,7 @@ def test_compute_reaction_rates_with_environments(spacetype) :
         system_dict["space"] = {"w":2, "environments":[0, 1]}
 
     system = rdsystem_from_dict(system_dict)
-    
+
     rates = strnkin.compute_reaction_rates(
         system = system,
         reaction = system.network.reactions[0],
@@ -158,10 +158,10 @@ def test_compute_reaction_rates_with_environments(spacetype) :
     assert rates[0].convert("molecule/s").value == 1
     assert rates[1].convert("molecule/s").value == 1
 
-def test_compute_reaction_rates_in_inhomogeneous_state(spacetype) :
+def _test_compute_reaction_rates_in_inhomogeneous_state(spacetype) :
     #checks that reaction env are taken into account
 
-    state = UnitArray([1, 2, 3, 4], "molecule")        
+    state = UnitArray([1, 2, 3, 4], "molecule")
     system_dict = {
         "network" : {
             "species" : [
@@ -181,9 +181,9 @@ def test_compute_reaction_rates_in_inhomogeneous_state(spacetype) :
             "edges" : []}
     else :
[…]
