### 2024-05-21 7162479
Update README.rst
update Testing section
