### 2024-05-21 52f0eb4
Update references.bib
Adding pytest reference
