+++ paper/references.bib
@@ -190,3 +190,11 @@
 	year = {1997},
 	pages = {1145--1164}
   	}
+
+@misc{refpytest,
+	title = "pytest 8.2.1",
+	author = {Krekel, Holger and Oliveira, Bruno and Pfannschmidt, Ronny and Bruynooghe, Floris and Laugher, Brianna and Bruhin, Florian},
+	year = {2004},
+	url = {https://github.com/pytest-dev/pytest},
+	note = {\url{https://github.com/pytest-dev/pytest} (Accessed online on May 21, 2024)}
+	}
