+++ paper/references.bib
@@ -163,7 +163,7 @@
 	title = {Building a reaction diffusion system and performing simulations. Documentation for Strengths},
 	month = may,
 	year = 2024,
-	note = {Accessed online on the 8th of may, 2024},
+	note = {Accessed online on May 8, 2024},
 	url = {https://strengths.readthedocs.io/en/latest/building_and_simulating_rds.html}
 	}
 
@@ -192,7 +192,7 @@
   	}
 
 @misc{refpytest,
-	title = "pytest 8.2.1",
+	title = {{pytest} 8.2.1},
 	author = {Krekel, Holger and Oliveira, Bruno and Pfannschmidt, Ronny and Bruynooghe, Floris and Laugher, Brianna and Bruhin, Florian},
 	year = {2004},
 	url = {https://github.com/pytest-dev/pytest},
