+++ paper/references.bib
@@ -196,5 +196,5 @@
 	author = {Krekel, Holger and Oliveira, Bruno and Pfannschmidt, Ronny and Bruynooghe, Floris and Laugher, Brianna and Bruhin, Florian},
 	year = {2004},
 	url = {https://github.com/pytest-dev/pytest},
-	note = {\url{https://github.com/pytest-dev/pytest} (Accessed online on May 21, 2024)}
+	note = {Accessed online on May 21, 2024}
 	}
