### 2024-05-21 8990757
Update building_and_simulating_rds.rst
fixing typos
