+++ documentation/building_and_simulating_rds.rst
@@ -1,7 +1,7 @@
-Building a reaction diffusion system and performing simulations
+Building a reaction-diffusion system and performing simulations
 ===============================================================
 
-This section illustrate with examples how to build and simulate reaction-diffusion systems
+This section illustrates with examples how to build and simulate reaction-diffusion systems
 using the different features proposed by Strengths.
 For more information about defining specific initial conditions in a system, or processing the simulation output,
 please refer to the dedicated sections.
@@ -9,7 +9,7 @@ please refer to the dedicated sections.
 A simple example without diffusion
 ----------------------------------
 
-Let us consider a simple simulation scenario :
+Let us consider a simple simulation scenario:
 
 We have a simple reaction network, with two species, *A* and *B*,
 and one reaction that goes as
@@ -33,12 +33,12 @@ and with initial densities for *A*, *B* and *C* of
   [B]_{t0} = 2 \textrm{ µM}\\
   [C]_{t0} = 0 \textrm{ µM}
 
-we want to simulate the time trajectory of this system using the Euler method with a time step of 1 ms,
-sampling the sate every 10 ms until 10s (in siumlation time).
+We want to simulate the time trajectory of this system using the Euler method with a time step of 1 ms,
+sampling the state every 10 ms until 10 s (in simulation time).
 
 Eventually, we want to plot the trajectories of *A*, *B* and *C* in one single figure.
 
-This can all be done with the following script :
+This can all be done with the following script:
 
 **system.json** :
 
@@ -81,7 +81,7 @@ This can all be done with the following script :
 .. image:: traj1.png
   :align: center
 
-let us explain in detail how this example works :
+let us explain in detail how this example works:
 
 .. code:: python
 
@@ -89,11 +89,11 @@ let us explain in detail how this example works :
   import numpy as np
   import strengths.plot as strnplt
 
-first, we imported 3 modules here : strengths (as strn), numpy (as np) and the strengths.plot subodule.
+First, we imported 3 modules here: strengths (as strn), numpy (`NumPy <https://numpy.org/>`_) (as np) and the strengths.plot subodule.
 strengths is the Strengths module, which contains everything required to perform simulations.
-numpy import is faculatative, it's used in the example for the linspace function, useful to define the sampling times.
-the strenghts.plot submodule is also faculatative. It bring conveinient function for trajectory plotting, working as a wrapper around
-matplotlib calls.
+the numpy import is optional, it's used in the example for the `linspace <https://numpy.org/doc/stable/reference/generated/numpy.linspace.html#numpy-linspace>`_ function, useful to define the sampling times.
+the strenghts.plot submodule is also optional. It brings convenient functions for trajectory plotting, working as a wrapper around
+the `Matplotlib package <https://matplotlib.org/>`_.
 
 .. code:: json
 
@@ -114,21 +114,21 @@ matplotlib calls.
   }
 
 Next, we define a reaction-diffusion system with a dictionary, in a JSON file (*system.json*).
-the dictionary defines two essencial components of the system :
+the dictionary defines two essential components of the system :
 
 * First, the reaction-diffusion "network", which tells
   which species compose the system, what their properties are and how they interact with each other. More specifically,
-  the list of the different species, whith their label/name and expected initial concentration, and a list of the different reactions,
-  with their stoechiometric equation and forward and backward rates.
+  the list of the different species, with their label/name and expected initial concentration, and a list of the different reactions,
+  with their stoichiometric equation and forward and backward rates.
 * Second, the system "space", which describes the reaction-diffusion grid in which the different species should evolve.
[…]
