+++ documentation/typechecking.rst
@@ -1,17 +1,0 @@
-Type Checking
-=============
-
-.. autofunction:: strengths.isnone
-
-.. autofunction:: strengths.isnumber
-
-.. autofunction:: strengths.isarray
-
-.. autofunction:: strengths.isdict
-
-.. autofunction:: strengths.isstr
-
-References
-----------
-
-.. [#numpy_ndarray] Numpy Developers. numpy 1.25 API reference : numpy.ndarray. (consulted on september 04, 2023). https://numpy.org/doc/stable/reference/generated/numpy.ndarray.html
