### 2024-05-21 a580c2d
Update index.rst
increasing the depth of the table of contents
