+++ paper/references.bib
@@ -160,10 +160,10 @@
 
 @misc{Strengths_doc,
 	author = {Fillion, Thibault and Piazza, Francesco},
-	title = {Building a reaction diffusion system and performing simulations. Documentation for Strengths},
+	title = {Building a reaction-diffusion system and performing simulations. Documentation for Strengths},
 	month = may,
 	year = 2024,
-	note = {Accessed online on May 8, 2024},
+	note = {Accessed online on May 21, 2024},
 	url = {https://strengths.readthedocs.io/en/latest/building_and_simulating_rds.html}
 	}
 
