### 2024-05-21 c11a828
Update building_strengths_from_source.rst
typos
