### 2024-05-21 ed94e48
Update arrays_and_numbers.rst
