### 2024-05-21 d77c485
Update arrays_and_numbers.rst
