+++ documentation/apiref.rst
@@ -14,7 +14,7 @@ Here are some things you might find interesting.
 Core features
 -------------
 
-Core features are those essential for reaction diffusion system building and simulation.
+Core features are those essential for reaction-diffusion system building and simulation.
 
 Directly imported
 ^^^^^^^^^^^^^^^^^
@@ -34,8 +34,8 @@ This is the part of the interface that is (mostly) directly accessible using "im
 Not directly imported
 ^^^^^^^^^^^^^^^^^^^^^
 
-This is the part of the interface that is not directly accessible from "import strengths".
-You will need to import those specific modules in order to access their features:
+This is the interface that is not directly accessible from "import strengths".
+You will need to import those specific modules to access their features:
 
 .. toctree::
   rdengine
@@ -46,7 +46,7 @@ Side features
 -------------
 
 Side features cover other aspects such as plots, and their corresponding submodules
-must be imported in order to access them:
+must be imported to access them:
 
 .. toctree::
   plot
