### 2024-05-21 4e9df3a
Update dependencies.rst
