### 2024-05-21 cb19382
Update fundamentals.rst
