+++ documentation/indexing.rst
@@ -3,11 +3,11 @@ Array indexing
 
 Multidimensional data is most of the time stored as unidimensional arrays.
 There are usually dedicated functions helping to access elements by coordinates,
-however it may still be of interest to understand how data are stored in the first place.
+however, it may still be of interest to understand how data are stored in the first place.
 This is what this section will do, by explaining how to compute the unidimensional array index
 from coordinates.
 
-Let N, w, h and d be respectively the number of species in a system, and the system width, height and depth in cell.
+Let N, w, h and d be respectively the number of species in a system, and the system width, height and depth in cells.
 
 Space cell index
 ----------------
