### 2024-05-21 9d9c985
Update rdnetwork.rst
