+++ paper/references.bib
@@ -160,7 +160,7 @@
 
 @misc{Strengths_doc,
 	author = {Fillion, Thibault and Piazza, Francesco},
-	title = {Building a reaction-diffusion system and performing simulations. Documentation for Strengths},
+	title = {Building and simulating a reaction-diffusion system. Documentation for Strengths},
 	month = may,
 	year = 2024,
 	note = {Accessed online on May 21, 2024},
