### 2024-05-21 4ff8a6f
Update rdscript.rst
