### 2024-05-21 1161b60
Update setting_up_initial_conditions.rst
