### 2024-05-21 ed5eb36
Update using_strengths.rst
