+++ documentation/using_strengths.rst
@@ -1,13 +1,13 @@
 Using Strengths
 ===============
 
-The following sections try to explain how to deal with various aspect of the Strengths package.
+The following sections try to explain how to deal with various aspects of the Strengths package.
 Each section is quite independent from the order and you can be approached in any order.
 
 The first section explains how to install and import the package.
 The second section shows how to build reaction-diffusion systems and perform system trajectory simulations.
-The third section explain how the UnitValue and UnitArray classes, which represent physical quantities, work.
-The Fourth section deal with simulations engines, explaining how they work and which one are available.
+The third section explains how the UnitValue and UnitArray classes, which represent physical quantities, work.
+The Fourth section deals with simulation engines, explaining how they work and which ones are available.
 The Last section just tells which external packages are needed to have strengths running.
 
 .. toctree::
