### 2024-05-21 93c939e
Update working_with_output.rst
