### 2024-05-21 e107ecc
Update doc and pytest support (#21)
* Update test_kinetics.py

* Add files via upload

Tests now rely on pytest.
run_all_tests.py is still here, but simply calls pytest.main() now.

* Update README.rst

update Testing section

* Update README.rst

* Update paper.md

adding a reference to pytest

* Update references.bib

Adding pytest reference

* Update references.bib

* Update references.bib

* Update README.rst

* Update building_and_simulating_rds.rst

fixing typos

* Delete documentation/typechecking.rst

* Update index.rst

increasing the depth of the table of contents

* Update reference
[…]
