+++ .github/workflows/build_wheels.yml
@@ -15,7 +15,7 @@ jobs:
       - uses: actions/checkout@v4
 
       - name: Build wheels
-        uses: pypa/cibuildwheel@v2.17.0
+        uses: pypa/cibuildwheel@v2.18.1
         # env:
         #   CIBW_SOME_OPTION: value
         #    ...
