### 2024-05-21 #6495 @ThibaultFillion
Hello @Nikoleta-v3,

Thank you very much for your corrections and suggestions.
We tried to implement all of them.
Here is a summary of the changes made:

* The paper branch has been merged with main, so that the paper and example simulation scripts lies next to the last (to be archived) version.
* The typo you pointed out in the paper has been fixed.
* pytest is a nice package, thank you for recommending it. The unit testing scripts in the "tests" directory were slightly adapted to work with pytest. For now, it still requires for the tests to be run from inside the "tests" directory, but it may change in the future.
* As a consequence, a reference to pytest has been added in the paper.
* The "Testing" section of "README.md" has also been updated.
* Thank you for pointing out the problem with the "Using Strengths" section.
The depth of the table of contents has been increased, allowing to
access the different sections and subsection from the lateral menu.
* We went through the documentation to fix more typos.

Since there were changes made in the tests directory, which is part of the distribution on PyPi, the version has been incremented to 0.0.16.
If everything is fine with this version, we will upload this version on PyPi and archive the repository on Zenodo, then proceed as you asked.
