### 2024-05-22 #6495 @Nikoleta-v3
Thank you for all the changes @ThibaultFillion! Everything looks great on my end.

> If everything is fine with this version, we will upload this version to PyPi and archive the repository on Zenodo, then proceed as you asked.

Yes, please go ahead!
