### 2024-05-22 #6495 @ThibaultFillion
@Nikoleta-v3 Perfect thank you, we are on it!

Also, @parikshitbajpai, since you contributed with commits, should i
add you to the Contributor section on Zenodo?

Otherwise, do i have to mention you somewhere on Zenodo as editor and reviewers @Nikoleta-v3, @parikshitbajpai and @jakryd?
