+++ paper/references.bib
@@ -135,7 +135,7 @@
 @book{Milo2015,
 	title={Cell biology by the numbers},
 	author={Milo, Ron and Phillips, Rob},
-	pages={258},
+	pages={213},
 	edition={1},
 	month={12},
 	year={2015},
