### 2024-05-23 c2473d5
Update paper.md
fixing "$\tau$ leap" to "$\tau$-leap"
