### 2024-05-23 #6495 @Nikoleta-v3
> Otherwise, do i have to mention you somewhere on Zenodo as editor and reviewers @Nikoleta-v3, @parikshitbajpai and @jakryd?

No, no need. We will appear on the paper, but thank you for asking 😄
