### 2024-05-23 #6495 @parikshitbajpai
> Also, @parikshitbajpai, since you contributed with commits, should i
>
> add you to the Contributor section on Zenodo?

No, that's not needed. I only made very minor changes to documentation as part of the review.
