### 2024-05-23 #6495 @ThibaultFillion
Thank you for your answers, @Nikoleta-v3  and @parikshitbajpai.

We just made two more edits before creating the final Zenodo archive:

* We fixed a typo in the paper: "$\tau$ leap" -> "$\tau$-leap" in the caption of the figure 3.
* We corrected the page number for the reference [Milo2016] in "references.bib".

Due to some difficulties, there are actually 3 archives on Zenodo.
The right one is the last one, with the same title than the paper:

  "STReNGTHS, a Python package to model and simulate reaction-diffusion systems".

It is labelled:

  v0.0.16.2  STReNGTHS, a Python package to model and simulate reaction-diffusion systems

and has the doi:

  10.5281/zenodo.11261599

Here is its url:

  https://zenodo.org/records/11261599

and the corresponding GitHub release is

  https://github.com/ThibaultFillion/strengths/releases/tag/v0.0.16.2

the wrong archives have labels starting with:

  v0.0.16 [to-be-deleted]

We will release the package on PyPi (later today, hopefully) as the version 0.0.16.
