### 2024-05-24 #6495 @Nikoleta-v3
Thank you @ThibaultFillion!

> We will release the package on PyPi (later today, hopefully) as the version 0.0.16.

If I remember correctly PyPi uses tag releases so it will be version v0.0.16.2 🤔 I could be wrong.
