### 2024-05-24 #6495 @ThibaultFillion
Hello @Nikoleta-v3,

Thank you for mentioning it.
We just uploaded the package on PyPI:

https://pypi.org/project/strengths/

However, the version number is 0.0.16.
Maybe it is because we are not uploading the package through GitHub.
Since the version number is only mentioned in pyproject.toml,
I assume this is where PiPI get the information from in this case.
