### 2024-05-24 #6495 @Nikoleta-v3
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1021/j100540a008 is OK
> - 10.1103/PhysRevE.71.041103 is OK
> - 10.1063/1.1378322 is OK
> - 10.1371/journal.pcbi.1006830 is OK
> - 10.1093/bioinformatics/bti431 is OK
> - 10.1049/sb:20045021 is OK
> - 10.1093/bioinformatics/btw469 is OK
> - 10.3389/neuro.11.015.2009 is OK
> - 10.1038/s41586-020-2649-2 is OK
> - 10.1109/MCSE.2007.55 is OK
> - 10.1201/9780429258770 is OK
> - 10.1091/mbc.E15-04-0186 is OK
> - 10.1016/S1468-1218(03)00020-8 is OK
> - 10.1007/BF02181277 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Building and simulating a reaction-diffusion syste...
> - No DOI given, and none found for title: pytest 8.2.1
>
> INVALID DOIs
>
> - None

Both of these references are online resources without DOIs. The authors have cited them correctly using the URLs of the resources and other necessary information.
