### 2024-05-24 #6495 @ThibaultFillion
@Nikoleta-v3 Thank you for recommending the paper for acceptance!
We just corrected the metadata of the Zenodo archive with the correct license (MIT License).
