### 2024-05-25 #6495 @Kevin-Mattheus-Moerman
- [ ] The version tag listed for this review should both match [one of the tags for your repository](https://github.com/ThibaultFillion/strengths/tags), as well as match the version listed on the ZENODO archive. Currently you have 6 versions on ZENODO but I cannot see one for `v0.0.16.2`. Should you edit the ZENODO listed version tag? Or should be edit the one listed for this review to be the latest archived version/tag e.g. `v0.0.16.5`, please advise.

![image](https://github.com/openjournals/joss-reviews/assets/8392709/590e3230-a244-416c-ad6f-e62bc7707953)
