### 2024-05-25 5769d8d
Update paper.md
fixing typos
