+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography : references.bib
 # Summary
 
 STReNGTHS is an open-source Python package that provides a simple and intuitive
-interface for designing models of discrete 3D heterogeneous reaction-diffusion systems and simulating their trajectories. Different algorithms are available, both stochastic (exact or approximate  solutions of the associated
+interface for designing models of discrete 3D heterogeneous reaction-diffusion systems and simulating their trajectories. Different algorithms are available, both stochastic (exact or approximate solutions of the associated
 master equation) and deterministic (numerical solutions of the corresponding rate equations).
 The acronym stands for "Simulation and modeling Tool for REaction-diffusion Networks in Graphs and Tridimensional Heterogeneous Systems" (STReNGTHS). The simulation algorithms are interfaced through a general abstract interface, which makes it easy to extend STReNGTHS with new algorithms and other features.  It is implemented in Python (standard library,
 Numpy [@harris_numpy_2020] and Matplotlib [@hunter_matplotlib_2007], as well as pytest [@refpytest] for unit testing)
@@ -40,7 +40,7 @@ organelles, stress granules, etc) and driven far from equilibrium by highly regu
 such as nucleotides (ATP/GTP), amino acids and different
 ions. In order to understand the properties of such reaction-diffusion
 networks, and especially their role in macroscopic *emergent*
-phenomena, convenient, reliable and efficient modeling and simulation tools supporting heterogeneous systems are necessary. Moreover, the choice of the simulation algorithm to be used may depend on the system : deterministic approaches such as ODE integration (rate equations) are effective, but inappropriate for systems that are sensitive to fluctuations,
+phenomena, convenient, reliable and efficient modeling and simulation tools supporting heterogeneous systems are necessary. Moreover, the choice of the simulation algorithm to be used may depend on the system: deterministic approaches such as ODE integration (rate equations) are effective, but inappropriate for systems that are sensitive to fluctuations,
 such as systems that operate with species present at low copy numbers
 (e.g. certain enzymes, many mRNA species) and/or
 in tiny reaction volumes (i.e. within mitochondria),
@@ -56,7 +56,7 @@ the *RDSystem* class. This defines a set of coupled chemical transformations
 that is supplemented with a spatial distribution of
 chemical species, whose features describe
 where and how fast each molecule shall diffuse.
-The system space is discrete, and consists of a 3D mesh of individual
+The system space is discrete and consists of a 3D mesh of individual
 volume elements, which we refer to as *cells*.
 It can be either a regular grid of cubic cells with
 uniform volumes, or an arbitrary network of cells with different volumes,
@@ -64,12 +64,12 @@ which can be obtained by coarse-graining a mesh grid.
 
 In order to account for systems with different compartments,
 STReNGTHS implements the system of reaction-diffusion environments,
-which allow the user to define different types of cells (referred to as environments)
+which allows the user to define different types of cells (referred to as environments)
 with specific reactive and diffusive properties. Many
-properties, such as the initial density of species, diffusion coefficients, or reactions
+properties, such as the initial density of species, diffusion coefficients, or reaction
 occurrence, can be defined environment-wise.
 
-Importantly, species can be *chemostatted*, i.e. kept at
+Importantly, species can be *chemostated*, i.e. kept at
 a fixed, prescribed concentration during the simulation,
[…]
