### 2024-05-25 3f552b9
Update paper.md
typo
