### 2024-05-25 fd38361
Update paper.md
typos + slight reformulation of one paragraph
