### 2024-05-25 #6495 @ThibaultFillion
Hello @Kevin-Mattheus-Moerman.

We would like to switch to the version v0.0.16.6.
Here is the corresponding Zenodo archive:

https://zenodo.org/records/11300235

The corresponding DOI is:

DOI: 10.5281/zenodo.11300235

Compared to the previous versions, we made fixed some typos and made a minor change in a paragraph (commit fd38361 in the main branch of the software repository), and rectified author affiliations.

Also, we would like to slightly change the title, previously:

"STReNGTHS, a Python package to model and simulate reaction-diffusion systems"

to:

"STReNGTHS, a Python package to model and simulate complex reaction-diffusion systems"

Sorry for all those last-minute changes.
