### 2024-05-27 #6495 @ThibaultFillion
Great news! Thank you @Kevin-Mattheus-Moerman, @Nikoleta-v3, @parikshitbajpai and @jakryd!
It was a pleasure publishing in JOSS!
